% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acov.R
\name{acov}
\alias{acov}
\title{Calculate the Asymptotic Variance}
\usage{
acov(D0, prm, est = c("same", "me", "mle"))
}
\arguments{
\item{D0}{A subclass of \code{Distribution}. The distribution family of interest.}

\item{prm}{A list containing three elements (name, pos, val). See Details.}

\item{est}{character. The estimator of interest. Can be a vector.}
}
\value{
A data.frame with columns named "Parameter", "Estimator", and
"Value".
}
\description{
This function calculates the asymptotic variance of an estimator as a
function of a single parameter, keeping the others constant. See Details.
}
\examples{
\dontrun{
# Distribution
d_beta <- Beta(shape1 = 1, shape2 = 1.5)

# Simulation
set.seed(2)
x <- r(d_beta)(50)

# Estimation
me(x, "Beta")
same(x, "Beta")
mle(x, "Beta")

# Asymptotic Covariance Matrix
acov_me(d_beta)
acov_same(d_beta)
acov_mle(d_beta)
}
}
