% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{acov_mle}
\alias{acov_mle}
\alias{acov_mle,Beta-method}
\alias{acov_mle,Gammad-method}
\alias{acov_mle,Dirichlet-method}
\alias{acov_mle,MGamma-method}
\title{Asymptotic covariance matrix of estimator}
\usage{
acov_mle(distr, ...)

\S4method{acov_mle}{Beta}(distr)

\S4method{acov_mle}{Gammad}(distr)

\S4method{acov_mle}{Dirichlet}(distr)

\S4method{acov_mle}{MGamma}(distr)
}
\arguments{
\item{distr}{A subclass of \code{Distribution}. The distribution of interest.}

\item{...}{extra arguments.}
}
\value{
matrix. The asymptotic covariance matrix of the estimator.
}
\description{
Asymptotic covariance matrix of estimator
}
