% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same.R
\name{acov_same}
\alias{acov_same}
\alias{acov_same,Beta-method}
\alias{acov_same,Gammad-method}
\alias{acov_same,Dirichlet-method}
\alias{acov_same,MGamma-method}
\title{Asymptotic covariance matrix of estimator}
\usage{
acov_same(distr, ...)

\S4method{acov_same}{Beta}(distr)

\S4method{acov_same}{Gammad}(distr)

\S4method{acov_same}{Dirichlet}(distr)

\S4method{acov_same}{MGamma}(distr)
}
\arguments{
\item{distr}{A subclass of \code{Distribution}. The distribution of interest.}

\item{...}{extra arguments.}
}
\value{
matrix. The asymptotic covariance matrix of the estimator.
}
\description{
Asymptotic covariance matrix of estimator
}
