% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_generics.R, R/02_Beta.R, R/02_Binom.R,
%   R/02_Dirichlet.R, R/02_Exp.R, R/02_Gammad.R, R/02_MGamma.R, R/02_Norm.R,
%   R/02_Pois.R
\name{avar}
\alias{avar}
\alias{vbeta}
\alias{vbinom}
\alias{vdirich}
\alias{vexp}
\alias{vgamma}
\alias{vmgamma}
\alias{vnorm}
\alias{vpois}
\title{Asymptotic Variance}
\usage{
avar(distr, type, ...)

vbeta(shape1, shape2, type = "mle")

vbinom(size, prob, type = "mle")

vdirich(shape, type = "mle")

vexp(rate, type = "mle")

vgamma(shape, scale, type = "mle")

vmgamma(shape, scale, type = "mle")

vnorm(mean, sd, type = "mle")

vpois(lambda, type = "mle")
}
\arguments{
\item{distr}{A subclass of \code{Distribution}. The distribution family assumed.}

\item{type}{character, case ignored. The estimator type (mle, me, or same).}

\item{...}{extra arguments.}

\item{shape1, shape2, size, prob, shape, rate, scale, mean, sd, lambda}{numeric.
Distribution parameters.}
}
\value{
A named matrix. The asymptotic covariance matrix of the estimator.
}
\description{
Calculates the asymptotic variance (or variance - covariance matrix in the
multidimensional case) of an estimator, given a specified family of
distributions and the true parameter values.
}
\examples{
\dontrun{
# -------------------------------------------
# Beta Distribution Example
# -------------------------------------------

# Simulation

set.seed(1)
x <- rbeta(100, shape1, shape2)
D <- Beta(shape1, shape2)

# Likelihood - The ll Functions

llbeta(x, shape1, shape2)
ll(x, c(shape1, shape2), D)
ll(x, c(shape1, shape2), "beta")

# Point Estimation - The e Functions

ebeta(x, type = "mle")
ebeta(x, type = "me")
ebeta(x, type = "same")

mle(x, D)
me(x, D)
same(x, D)

estim(x, D, type = "mle")

# Asymptotic Variance - The v Functions

vbeta(shape1, shape2, type = "mle")
vbeta(shape1, shape2, type = "me")
vbeta(shape1, shape2, type = "same")

avar_mle(D)
avar_me(D)
avar_same(D)

avar(D, type = "mle")
}
}
\seealso{
\link{avar_mle}, \link{avar_me}, \link{avar_same}
}
