% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Bern.R
\name{Bern}
\alias{Bern}
\alias{d,Bern-method}
\alias{p,Bern-method}
\alias{q2,Bern-method}
\alias{r,Bern-method}
\alias{mean,Bern-method}
\alias{median,Bern-method}
\alias{mode,Bern-method}
\alias{var,Bern-method}
\alias{sd,Bern-method}
\alias{skew,Bern-method}
\alias{kurt,Bern-method}
\alias{entro,Bern-method}
\alias{finf,Bern-method}
\alias{ll,numeric,numeric,Bern-method}
\alias{mle,numeric,Bern-method}
\alias{me,numeric,Bern-method}
\alias{avar_mle,Bern-method}
\alias{avar_me,Bern-method}
\title{Bernoulli Distribution}
\usage{
Bern(prob = 0.5)

\S4method{d}{Bern}(x)

\S4method{p}{Bern}(x)

\S4method{q2}{Bern}(x)

\S4method{r}{Bern}(x)

\S4method{mean}{Bern}(x)

\S4method{median}{Bern}(x)

\S4method{mode}{Bern}(x)

\S4method{var}{Bern}(x)

\S4method{sd}{Bern}(x)

\S4method{skew}{Bern}(x)

\S4method{kurt}{Bern}(x)

\S4method{entro}{Bern}(x)

\S4method{finf}{Bern}(x)

\S4method{ll}{numeric,numeric,Bern}(x, prm, distr)

\S4method{mle}{numeric,Bern}(x, distr)

\S4method{me}{numeric,Bern}(x, distr)

\S4method{avar_mle}{Bern}(distr)

\S4method{avar_me}{Bern}(distr)
}
\arguments{
\item{x}{an object of class \code{Bern}. If the function also has a \code{distr}
argument, \code{x} is a numeric vector, a sample of observations.}

\item{prm, prob}{numeric. The distribution parameter.}

\item{distr}{an object of class \code{Bern}.}
}
\description{
Bernoulli Distribution
}
