% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Beta.R
\name{Beta}
\alias{Beta}
\alias{d,Beta-method}
\alias{p,Beta-method}
\alias{q2,Beta-method}
\alias{r,Beta-method}
\alias{mean,Beta-method}
\alias{median,Beta-method}
\alias{mode,Beta-method}
\alias{var,Beta-method}
\alias{sd,Beta-method}
\alias{skew,Beta-method}
\alias{kurt,Beta-method}
\alias{entro,Beta-method}
\alias{finf,Beta-method}
\alias{llbeta}
\alias{ll,numeric,numeric,Beta-method}
\alias{mle,numeric,Beta-method}
\alias{me,numeric,Beta-method}
\alias{same,numeric,Beta-method}
\alias{vbeta}
\alias{avar_mle,Beta-method}
\alias{avar_me,Beta-method}
\alias{avar_same,Beta-method}
\title{Beta Distribution}
\usage{
Beta(shape1 = 1, shape2 = 1, ncp = 0)

\S4method{d}{Beta}(x)

\S4method{p}{Beta}(x)

\S4method{q2}{Beta}(x)

\S4method{r}{Beta}(x)

\S4method{mean}{Beta}(x)

\S4method{median}{Beta}(x)

\S4method{mode}{Beta}(x)

\S4method{var}{Beta}(x)

\S4method{sd}{Beta}(x)

\S4method{skew}{Beta}(x)

\S4method{kurt}{Beta}(x)

\S4method{entro}{Beta}(x)

\S4method{finf}{Beta}(x)

llbeta(x, shape1, shape2)

\S4method{ll}{numeric,numeric,Beta}(x, prm, distr)

\S4method{mle}{numeric,Beta}(x, distr, par0 = "same", method = "L-BFGS-B", lower = 1e-05, upper = Inf)

\S4method{me}{numeric,Beta}(x, distr)

\S4method{same}{numeric,Beta}(x, distr)

vbeta(shape1, shape2, type = "mle")

\S4method{avar_mle}{Beta}(distr)

\S4method{avar_me}{Beta}(distr)

\S4method{avar_same}{Beta}(distr)
}
\arguments{
\item{shape1, shape2, ncp}{numeric. The distribution parameters.}

\item{x}{an object of class \code{Beta}. If the function also has a \code{distr}
argument, \code{x} is a numeric vector, a sample of observations.}

\item{prm}{numeric. A vector including the distribution parameters.}

\item{distr}{an object of class \code{Beta}.}

\item{par0, method, lower, upper}{arguments passed to optim.}

\item{type}{character, case ignored. The estimator type (mle, me, or same).}
}
\description{
Beta Distribution
}
