% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Binom.R
\name{Binom}
\alias{Binom}
\alias{d,Binom-method}
\alias{p,Binom-method}
\alias{q2,Binom-method}
\alias{r,Binom-method}
\alias{mean,Binom-method}
\alias{var,Binom-method}
\alias{sd,Binom-method}
\alias{skew,Binom-method}
\alias{kurt,Binom-method}
\alias{finf,Binom-method}
\alias{ll,numeric,numeric,Binom-method}
\alias{mle,numeric,Binom-method}
\alias{me,numeric,Binom-method}
\alias{avar_mle,Binom-method}
\alias{avar_me,Binom-method}
\title{Binomial Distribution}
\usage{
Binom(size = 1, prob = 0.5)

\S4method{d}{Binom}(x)

\S4method{p}{Binom}(x)

\S4method{q2}{Binom}(x)

\S4method{r}{Binom}(x)

\S4method{mean}{Binom}(x)

\S4method{var}{Binom}(x)

\S4method{sd}{Binom}(x)

\S4method{skew}{Binom}(x)

\S4method{kurt}{Binom}(x)

\S4method{finf}{Binom}(x)

\S4method{ll}{numeric,numeric,Binom}(x, prm, distr)

\S4method{mle}{numeric,Binom}(x, distr)

\S4method{me}{numeric,Binom}(x, distr)

\S4method{avar_mle}{Binom}(distr)

\S4method{avar_me}{Binom}(distr)
}
\arguments{
\item{size, prob}{numeric. The distribution parameters.}

\item{x}{an object of class \code{Binom}. If the function also has a \code{distr}
argument, \code{x} is a numeric vector, a sample of observations.}

\item{prm}{numeric. A vector including the distribution parameters.}

\item{distr}{an object of class \code{Binom}.}
}
\description{
Binomial Distribution
}
