% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_calculus.R
\name{calculus}
\alias{calculus}
\alias{+,Norm,Norm-method}
\alias{+,numeric,Norm-method}
\alias{+,Norm,numeric-method}
\alias{-,Norm,Norm-method}
\alias{-,numeric,Norm-method}
\alias{-,Norm,numeric-method}
\alias{*,numeric,Norm-method}
\alias{*,Norm,numeric-method}
\alias{/,Norm,numeric-method}
\alias{sum,Norm,logical-method}
\alias{exp,Norm-method}
\title{Distribution Calculus}
\usage{
\S4method{+}{Norm,Norm}(e1, e2)

\S4method{+}{numeric,Norm}(e1, e2)

\S4method{+}{Norm,numeric}(e1, e2)

\S4method{-}{Norm,Norm}(e1, e2)

\S4method{-}{numeric,Norm}(e1, e2)

\S4method{-}{Norm,numeric}(e1, e2)

\S4method{*}{numeric,Norm}(e1, e2)

\S4method{*}{Norm,numeric}(e1, e2)

\S4method{/}{Norm,numeric}(e1, e2)

\S4method{sum}{Norm,logical}(x, ..., na.rm = FALSE)

\S4method{exp}{Norm}(x)
}
\arguments{
\item{x, e1, e2}{objects of subclass \code{Distribution}.}

\item{...}{extra arguments.}

\item{na.rm}{logical. Should missing values be removed?}
}
\description{
Distribution Calculus
}
