% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_generics.R
\name{d}
\alias{d}
\alias{p}
\alias{q2}
\alias{r}
\title{d, p, q, r}
\usage{
d(x, ...)

p(x, ...)

q2(x, ...)

r(x, ...)
}
\arguments{
\item{x}{an object of subclass \code{Distribution}.}

\item{...}{extra arguments.}
}
\value{
function.
}
\description{
d, p, q, r
}
\examples{
1 + 1
}
