% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dDirichlet}
\alias{dDirichlet}
\alias{rDirichlet}
\title{The Dirichlet Distribution}
\usage{
dDirichlet(x, shape, log = FALSE)

rDirichlet(n, shape)
}
\arguments{
\item{x}{numeric. The quantile vector.}

\item{shape}{numeric. The parameter vector.}

\item{log}{logical. If TRUE, probabilities p are given as log(p).}

\item{n}{numeric. The number of observations.}
}
\value{
\code{dDirichlet} returns the evaluated density function.
\code{rDirichlet} performs Monte-Carlo simulation.
}
\description{
The Dirichlet Distribution
}
\examples{
\dontrun{
# Classic R Stats Format
dDirichlet(c(0.3, 0.7), shape = c(2, 3))
set.seed(1)
rDirichlet(10, shape = c(2, 3))

# S4 Distribution Class
d_dirichlet <- Dirichlet(shape = c(2, 3))
d(d_dirichlet)(c(0.3, 0.7))
set.seed(1)
r(d_dirichlet)(10)
}
}
