% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Dir.R
\name{Dir}
\alias{Dir}
\alias{d,Dir,numeric-method}
\alias{r,Dir,numeric-method}
\alias{mean,Dir-method}
\alias{mode,Dir-method}
\alias{var,Dir-method}
\alias{entro,Dir-method}
\alias{finf,Dir-method}
\alias{lldirichlet}
\alias{ll,Dir,matrix-method}
\alias{edirichlet}
\alias{mle,Dir,matrix-method}
\alias{me,Dir,matrix-method}
\alias{same,Dir,matrix-method}
\alias{vdirichlet}
\alias{avar_mle,Dir-method}
\alias{avar_me,Dir-method}
\alias{avar_same,Dir-method}
\title{Dirichlet Distribution}
\usage{
Dir(alpha = c(1, 1))

\S4method{d}{Dir,numeric}(distr, x)

\S4method{r}{Dir,numeric}(distr, n)

\S4method{mean}{Dir}(x)

\S4method{mode}{Dir}(x)

\S4method{var}{Dir}(x)

\S4method{entro}{Dir}(x)

\S4method{finf}{Dir}(x)

lldirichlet(x, alpha)

\S4method{ll}{Dir,matrix}(distr, x)

edirichlet(x, type = "mle", ...)

\S4method{mle}{Dir,matrix}(distr, x, par0 = "same", method = "L-BFGS-B", lower = 1e-05, upper = Inf)

\S4method{me}{Dir,matrix}(distr, x)

\S4method{same}{Dir,matrix}(distr, x)

vdirichlet(alpha, type = "mle")

\S4method{avar_mle}{Dir}(distr)

\S4method{avar_me}{Dir}(distr)

\S4method{avar_same}{Dir}(distr)
}
\arguments{
\item{alpha}{numeric. The distribution parameters.}

\item{distr}{an object of class \code{Dir}.}

\item{x}{an object of class \code{Dir}. If the function also has a \code{distr}
argument, \code{x} is a numeric vector, a sample of observations.}

\item{n}{numeric. The sample size.}

\item{type}{character, case ignored. The estimator type (mle, me, or same).}

\item{...}{extra arguments.}

\item{par0, method, lower, upper}{arguments passed to optim.}
}
\value{
Each type of function returns a different type of object:
\itemize{
\item Distribution Functions: When supplied with one argument (\code{distr}), the
\code{d()}, \code{p()}, \code{q()}, \code{r()}, \code{ll()} functions return the density, cumulative
probability, quantile, random sample generator, and log-likelihood functions,
respectively. When supplied with both arguments (\code{distr} and \code{x}), they
evaluate the aforementioned functions directly.
\item Moments: Returns a numeric, either vector or matrix depending on the moment
and the distribution. The \code{moments()} function returns a list with all the
available methods.
\item Estimation: Returns a list. The estimator of the unknown parameters. Note
that in distribution families like the binomial, multinomial, and negative
binomial, the size is not returned, since it is considered known.
\item Variance: Returns a named matrix. The asymptotic covariance matrix of the
estimator.
}
}
\description{
Dirichlet Distribution
}
