% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Dirichlet.R
\docType{class}
\name{Dirichlet-class}
\alias{Dirichlet-class}
\alias{Dirichlet}
\title{Dirichlet Distribution S4 Class}
\value{
An object of class \code{Dirichlet}.
}
\description{
Dirichlet Distribution S4 Class
}
\section{Slots}{

\describe{
\item{\code{shape}}{numeric. The parameter vector.}
}}

\examples{
# Classic R Stats Format
ddirich(c(0.3, 0.7), shape = c(2, 3))
set.seed(1)
rdirich(10, shape = c(2, 3))

# S4 Distribution Class
library(distr)
D <- Dirichlet(shape = c(2, 3))
d(D)(c(0.3, 0.7))
set.seed(1)
r(D)(10)
}
