% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_generics.R
\name{DistrFunctionals}
\alias{DistrFunctionals}
\alias{d,Distribution,missing-method}
\alias{p,Distribution,missing-method}
\alias{qn,Distribution,missing-method}
\alias{r,Distribution,missing-method}
\alias{ll,Distribution,missing-method}
\alias{mle,Distribution,missing-method}
\alias{me,Distribution,missing-method}
\alias{same,Distribution,missing-method}
\title{Distribution Functionals}
\usage{
\S4method{d}{Distribution,missing}(distr, x, ...)

\S4method{p}{Distribution,missing}(distr, x, ...)

\S4method{qn}{Distribution,missing}(distr, x, ...)

\S4method{r}{Distribution,missing}(distr, n, ...)

\S4method{ll}{Distribution,missing}(distr, x, ...)

\S4method{mle}{Distribution,missing}(distr, x, ...)

\S4method{me}{Distribution,missing}(distr, x, ...)

\S4method{same}{Distribution,missing}(distr, x, ...)
}
\arguments{
\item{distr}{a \code{Distribution} object.}

\item{x, n}{missing. Agruments not supplied.}

\item{...}{extra arguments.}
}
\value{
When supplied with one argument, the \code{d()}, \code{p()}, \code{q()}, \code{r()} \code{ll()}
functions return the density, cumulative probability, quantile, random sample
generator, and log-likelihood functions, respectively.
}
\description{
A collection of S4 classes that provide a flexible and structured way to work
with probability distributions.
}
\details{
When \code{x} or \code{n} are missing, the methods return a function that takes as
input the missing argument, allowing the user to work with the function
object itself. See examples.
}
\examples{
# -----------------------------------------------------
# Beta Distribution Example
# -----------------------------------------------------

# Create the distribution
D <- Beta(3, 5)
x <- c(0.3, 0.8, 0.5)
n <- 5

# Density function
df <- d(D) ; df(x) # df is a function itself
d(D, x) # alternative way to use the function

# Distribution function
pf <- p(D) ; pf(x)
p(D, x)

# Inverse distribution function
qf <- qn(D) ; qf(x)
qn(D, x)

# Random Generator function
rf <- r(D) ; rf(n)
r(D, n)
}
\seealso{
\link{moments}, \link{loglikelihood}, \link{estimation}, \link{Bern},
\link{Beta}, \link{Binom}, \link{Cat}, \link{Cauchy}, \link{Chisq}, \link{Dir}, \link{Exp},
\link{Fisher}, \link{Gam}, \link{Geom}, \link{Laplace}, \link{Lnorm}, \link{Multigam}, \link{Multinom},
\link{Nbinom}, \link{Norm}, \link{Pois}, \link{Stud}, \link{Unif}, \link{Weib}
}
