% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_generics.R
\name{Distribution-class}
\alias{Distribution-class}
\title{Distribution S4 Classes}
\value{
An object of the respective class (\code{Bern}, \code{Norm}, etc.). All classes
are subclasses of the virtual class \code{Distribution}.
}
\description{
A collection of classes that provide a flexible and structured way to work
with probability distributions.
}
\examples{
# -----------------------------------------------------
# Distribution Calculus Example
# -----------------------------------------------------

# Normal location - scale transformation
x <- Norm(mean = 2, sd = 3)
y <- 3 * x + 1 # Norm(mean = 7, sd = 9)

# Addition of two independent Normal random variables
x1 <- Norm(mean = 1, sd = 3)
x2 <- Norm(mean = 2, sd = 4)
x3 <- x1 + x2 # Norm(mean = 3, sd = 5)
}
