% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dMGamma}
\alias{dMGamma}
\alias{rMGamma}
\title{The Multivariate Gamma Distribution}
\usage{
dMGamma(x, shape, scale, log = FALSE)

rMGamma(n, shape, scale)
}
\arguments{
\item{x}{numeric. The quantile vector.}

\item{shape}{numeric. The shape parameter vector.}

\item{scale}{numeric. The scale parameter vector.}

\item{log}{logical. If TRUE, probabilities p are given as log(p).}

\item{n}{numeric. The number of observations.}
}
\value{
\code{dMGamma} returns the evaluated density function.
\code{rMGamma} performs Monte-Carlo simulation.
}
\description{
The Multivariate Gamma Distribution
}
\examples{
\dontrun{
# Classic R Stats Format
dMGamma(c(4, 6), shape = c(2, 3), scale = 2)
set.seed(1)
rMGamma(10, shape = c(2, 3), scale = 2)

# S4 Distribution Class
d_MGamma <- MGamma(shape = c(2, 3), scale = 2)
d(d_MGamma)(c(4, 6))
set.seed(1)
r(d_MGamma)(10)
}
}
