% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_generics.R, R/02_Beta.R, R/02_Binom.R,
%   R/02_Dirichlet.R, R/02_Exp.R, R/02_Gammad.R, R/02_MGamma.R, R/02_Norm.R,
%   R/02_Pois.R
\name{estim}
\alias{estim}
\alias{ebeta}
\alias{ebinom}
\alias{edirich}
\alias{eexp}
\alias{egamma}
\alias{emgamma}
\alias{enorm}
\alias{epois}
\title{Parameter Estimation}
\usage{
estim(x, distr, type = "mle", ...)

ebeta(x, type = "mle", ...)

ebinom(x, type = "mle", ...)

edirich(x, type = "mle", ...)

eexp(x, type = "mle", ...)

egamma(x, type = "mle", ...)

emgamma(x, type = "mle", ...)

enorm(x, type = "mle", ...)

epois(x, type = "mle", ...)
}
\arguments{
\item{x}{numeric. A sample under estimation.}

\item{distr}{A subclass of \code{Distribution}. The distribution family assumed.}

\item{type}{character, case ignored. The estimator type (mle, me, or same).}

\item{...}{extra arguments.}
}
\value{
numeric. The estimator produced by the sample.
}
\description{
Estimates the parameters of a random sample according to a
specified family of distributions.
}
\examples{
\dontrun{
# -------------------------------------------
# Beta Distribution Example
# -------------------------------------------

# Simulation

set.seed(1)
x <- rbeta(100, shape1, shape2)
D <- Beta(shape1, shape2)

# Likelihood - The ll Functions

llbeta(x, shape1, shape2)
ll(x, c(shape1, shape2), D)
ll(x, c(shape1, shape2), "beta")

# Point Estimation - The e Functions

ebeta(x, type = "mle")
ebeta(x, type = "me")
ebeta(x, type = "same")

mle(x, D)
me(x, D)
same(x, D)

estim(x, D, type = "mle")

# Asymptotic Variance - The v Functions

vbeta(shape1, shape2, type = "mle")
vbeta(shape1, shape2, type = "me")
vbeta(shape1, shape2, type = "same")

avar_mle(D)
avar_me(D)
avar_same(D)

avar(D, type = "mle")
}
}
\seealso{
\link{mle}, \link{me}, \link{same}
}
