% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators-package.R
\docType{package}
\name{estimators-package}
\alias{estimators}
\alias{estimators-package}
\title{estimators: Distribution Parameter Estimation Methods}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package provides estimators for parameters of common distribution families.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://thechibo.github.io/estimators/}
}

}
\author{
\strong{Maintainer}: Ioannis Oikonomidis \email{goikon@math.uoa.gr} (\href{https://orcid.org/0000-0001-8130-2104}{ORCID})

Authors:
\itemize{
  \item Samis Trevezas \email{strevezas@math.uoa.gr} (\href{https://orcid.org/0000-0003-2262-8299}{ORCID}) [thesis advisor]
}

}
\keyword{internal}
