% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Gam.R
\name{Gam}
\alias{Gam}
\alias{d,Gam,numeric-method}
\alias{p,Gam,numeric-method}
\alias{qn,Gam,numeric-method}
\alias{r,Gam,numeric-method}
\alias{mean,Gam-method}
\alias{median,Gam-method}
\alias{mode,Gam-method}
\alias{var,Gam-method}
\alias{sd,Gam-method}
\alias{skew,Gam-method}
\alias{kurt,Gam-method}
\alias{entro,Gam-method}
\alias{finf,Gam-method}
\alias{llgamma}
\alias{ll,Gam,numeric-method}
\alias{egamma}
\alias{mle,Gam,numeric-method}
\alias{me,Gam,numeric-method}
\alias{same,Gam,numeric-method}
\alias{vgamma}
\alias{avar_mle,Gam-method}
\alias{avar_me,Gam-method}
\alias{avar_same,Gam-method}
\title{Gamma Distribution}
\usage{
Gam(shape = 1, scale = 1)

\S4method{d}{Gam,numeric}(distr, x)

\S4method{p}{Gam,numeric}(distr, x)

\S4method{qn}{Gam,numeric}(distr, x)

\S4method{r}{Gam,numeric}(distr, n)

\S4method{mean}{Gam}(x)

\S4method{median}{Gam}(x)

\S4method{mode}{Gam}(x)

\S4method{var}{Gam}(x)

\S4method{sd}{Gam}(x)

\S4method{skew}{Gam}(x)

\S4method{kurt}{Gam}(x)

\S4method{entro}{Gam}(x)

\S4method{finf}{Gam}(x)

llgamma(x, shape, scale)

\S4method{ll}{Gam,numeric}(distr, x)

egamma(x, type = "mle", ...)

\S4method{mle}{Gam,numeric}(distr, x, par0 = "same", method = "L-BFGS-B", lower = 1e-05, upper = Inf)

\S4method{me}{Gam,numeric}(distr, x)

\S4method{same}{Gam,numeric}(distr, x)

vgamma(shape, scale, type = "mle")

\S4method{avar_mle}{Gam}(distr)

\S4method{avar_me}{Gam}(distr)

\S4method{avar_same}{Gam}(distr)
}
\arguments{
\item{shape, scale}{numeric. The distribution parameters.}

\item{distr}{an object of class \code{Gam}.}

\item{x}{an object of class \code{Gam}. If the function also has a \code{distr}
argument, \code{x} is a numeric vector, a sample of observations.}

\item{n}{numeric. The sample size.}

\item{type}{character, case ignored. The estimator type (mle, me, or same).}

\item{...}{extra arguments.}

\item{par0, method, lower, upper}{arguments passed to optim.}
}
\value{
Each type of function returns a different type of object:
\itemize{
\item Distribution Functions: When supplied with one argument (\code{distr}), the
\code{d()}, \code{p()}, \code{q()}, \code{r()}, \code{ll()} functions return the density, cumulative
probability, quantile, random sample generator, and log-likelihood functions,
respectively. When supplied with both arguments (\code{distr} and \code{x}), they
evaluate the aforementioned functions directly.
\item Moments: Returns a numeric, either vector or matrix depending on the moment
and the distribution. The \code{moments()} function returns a list with all the
available methods.
\item Estimation: Returns a list. The estimator of the unknown parameters. Note
that in distribution families like the binomial, multinomial, and negative
binomial, the size is not returned, since it is considered known.
\item Variance: Returns a named matrix. The asymptotic covariance matrix of the
estimator.
}
}
\description{
Gamma Distribution
}
