% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Geom.R
\name{Geom}
\alias{Geom}
\alias{d,Geom,numeric-method}
\alias{p,Geom,numeric-method}
\alias{qn,Geom,numeric-method}
\alias{r,Geom,numeric-method}
\alias{mean,Geom-method}
\alias{median,Geom-method}
\alias{mode,Geom-method}
\alias{var,Geom-method}
\alias{sd,Geom-method}
\alias{skew,Geom-method}
\alias{kurt,Geom-method}
\alias{entro,Geom-method}
\alias{finf,Geom-method}
\alias{llgeom}
\alias{ll,Geom,numeric-method}
\alias{egeom}
\alias{mle,Geom,numeric-method}
\alias{me,Geom,numeric-method}
\alias{vgeom}
\alias{avar_mle,Geom-method}
\alias{avar_me,Geom-method}
\title{Geometric Distribution}
\usage{
Geom(prob = 0.5)

\S4method{d}{Geom,numeric}(distr, x)

\S4method{p}{Geom,numeric}(distr, x)

\S4method{qn}{Geom,numeric}(distr, x)

\S4method{r}{Geom,numeric}(distr, n)

\S4method{mean}{Geom}(x)

\S4method{median}{Geom}(x)

\S4method{mode}{Geom}(x)

\S4method{var}{Geom}(x)

\S4method{sd}{Geom}(x)

\S4method{skew}{Geom}(x)

\S4method{kurt}{Geom}(x)

\S4method{entro}{Geom}(x)

\S4method{finf}{Geom}(x)

llgeom(x, prob)

\S4method{ll}{Geom,numeric}(distr, x)

egeom(x, type = "mle", ...)

\S4method{mle}{Geom,numeric}(distr, x)

\S4method{me}{Geom,numeric}(distr, x)

vgeom(prob, type = "mle")

\S4method{avar_mle}{Geom}(distr)

\S4method{avar_me}{Geom}(distr)
}
\arguments{
\item{prob}{numeric. The distribution parameters.}

\item{distr}{an object of class \code{Geom}.}

\item{x}{an object of class \code{Geom}. If the function also has a \code{distr}
argument, \code{x} is a numeric vector, a sample of observations.}

\item{n}{numeric. The sample size.}

\item{type}{character, case ignored. The estimator type (mle, me, or same).}

\item{...}{extra arguments.}
}
\value{
Each type of function returns a different type of object:
\itemize{
\item Distribution Functions: When supplied with one argument (\code{distr}), the
\code{d()}, \code{p()}, \code{q()}, \code{r()}, \code{ll()} functions return the density, cumulative
probability, quantile, random sample generator, and log-likelihood functions,
respectively. When supplied with both arguments (\code{distr} and \code{x}), they
evaluate the aforementioned functions directly.
\item Moments: Returns a numeric, either vector or matrix depending on the moment
and the distribution. The \code{moments()} function returns a list with all the
available methods.
\item Estimation: Returns a list. The estimator of the unknown parameters. Note
that in distribution families like the binomial, multinomial, and negative
binomial, the size is not returned, since it is considered known.
\item Variance: Returns a named matrix. The asymptotic covariance matrix of the
estimator.
}
}
\description{
Geometric Distribution
}
