% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Lnorm.R
\name{Lnorm}
\alias{Lnorm}
\alias{d,Lnorm,numeric-method}
\alias{p,Lnorm,numeric-method}
\alias{qn,Lnorm,numeric-method}
\alias{r,Lnorm,numeric-method}
\alias{mean,Lnorm-method}
\alias{median,Lnorm-method}
\alias{mode,Lnorm-method}
\alias{var,Lnorm-method}
\alias{sd,Lnorm-method}
\alias{skew,Lnorm-method}
\alias{kurt,Lnorm-method}
\alias{entro,Lnorm-method}
\alias{finf,Lnorm-method}
\alias{lllnorm}
\alias{ll,Lnorm,numeric-method}
\alias{elnorm}
\alias{mle,Lnorm,numeric-method}
\alias{me,Lnorm,numeric-method}
\alias{vlnorm}
\alias{avar_mle,Lnorm-method}
\alias{avar_me,Lnorm-method}
\title{Lnorm Distribution}
\usage{
Lnorm(meanlog = 0, sdlog = 1)

\S4method{d}{Lnorm,numeric}(distr, x)

\S4method{p}{Lnorm,numeric}(distr, x)

\S4method{qn}{Lnorm,numeric}(distr, x)

\S4method{r}{Lnorm,numeric}(distr, n)

\S4method{mean}{Lnorm}(x)

\S4method{median}{Lnorm}(x)

\S4method{mode}{Lnorm}(x)

\S4method{var}{Lnorm}(x)

\S4method{sd}{Lnorm}(x)

\S4method{skew}{Lnorm}(x)

\S4method{kurt}{Lnorm}(x)

\S4method{entro}{Lnorm}(x)

\S4method{finf}{Lnorm}(x)

lllnorm(x, meanlog, sdlog)

\S4method{ll}{Lnorm,numeric}(distr, x)

elnorm(x, type = "mle", ...)

\S4method{mle}{Lnorm,numeric}(distr, x)

\S4method{me}{Lnorm,numeric}(distr, x)

vlnorm(meanlog, sdlog, type = "mle")

\S4method{avar_mle}{Lnorm}(distr)

\S4method{avar_me}{Lnorm}(distr)
}
\arguments{
\item{meanlog, sdlog}{numeric. The distribution parameters.}

\item{distr}{an object of class \code{Lnorm}.}

\item{x}{an object of class \code{Lnorm}. If the function also has a \code{distr}
argument, \code{x} is a numeric vector, a sample of observations.}

\item{n}{numeric. The sample size.}

\item{type}{character, case ignored. The estimator type (mle, me, or same).}

\item{...}{extra arguments.}
}
\value{
Each type of function returns a different type of object:
\itemize{
\item Distribution Functions: When supplied with one argument (\code{distr}), the
\code{d()}, \code{p()}, \code{q()}, \code{r()}, \code{ll()} functions return the density, cumulative
probability, quantile, random sample generator, and log-likelihood functions,
respectively. When supplied with both arguments (\code{distr} and \code{x}), they
evaluate the aforementioned functions directly.
\item Moments: Returns a numeric, either vector or matrix depending on the moment
and the distribution. The \code{moments()} function returns a list with all the
available methods.
\item Estimation: Returns a list. The estimator of the unknown parameters. Note
that in distribution families like the binomial, multinomial, and negative
binomial, the size is not returned, since it is considered known.
\item Variance: Returns a named matrix. The asymptotic covariance matrix of the
estimator.
}
}
\description{
Lnorm Distribution
}
