% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.R
\name{me}
\alias{me}
\alias{me,ANY,character-method}
\alias{me,numeric,Beta-method}
\alias{me,matrix,Beta-method}
\alias{me,numeric,Gammad-method}
\alias{me,matrix,Gammad-method}
\alias{me,matrix,Dirichlet-method}
\alias{me,array,MGamma-method}
\title{Moment Estimator}
\usage{
me(x, distr, ...)

\S4method{me}{ANY,character}(x, distr, ...)

\S4method{me}{numeric,Beta}(x, distr)

\S4method{me}{matrix,Beta}(x, distr)

\S4method{me}{numeric,Gammad}(x, distr)

\S4method{me}{matrix,Gammad}(x, distr)

\S4method{me}{matrix,Dirichlet}(x, distr)

\S4method{me}{array,MGamma}(x, distr)
}
\arguments{
\item{x}{numeric. A sample under estimation.}

\item{distr}{A subclass of \code{Distribution}. The distribution of interest.}

\item{...}{extra arguments.}
}
\value{
numeric. The estimator produced by the sample.
}
\description{
Moment Estimator
}
