% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Calculate Metrics}
\usage{
metrics(
  D0,
  prm,
  obs = c(20, 50, 100),
  est = c("same", "me", "mle"),
  sam = 10000,
  seed = 1
)
}
\arguments{
\item{D0}{A subclass of \code{Distribution}. The distribution family of interest.}

\item{prm}{A list containing three elements (name, pos, val). See Details.}

\item{obs}{numeric. The size of each sample. Can be a vector.}

\item{est}{character. The estimator of interest. Can be a vector.}

\item{sam}{numeric. The number of Monte Carlo samples used to estimate the
metrics.}

\item{seed}{numeric. Passed to set.seed() for reproducibility.}
}
\value{
A data.frame with columns named "Parameter", "Observations",
"Estimator", "Metric", and "Value".
}
\description{
This function performs Monte Carlo simulations to estimate the main metrics
(bias, variance, and RMSE) characterizing the finite sample behavior of an
estimator. The function evaluates the metrics as a function of a single
parameter, keeping the other ones constant. See Details.
}
\details{
The distribution \code{D0} is used to specify an initial distribution. The list
\code{prm} contains details concerning a single parameter that is allowed to
change values. The quantity of interest is evaluated as a function of this
parameter.

Specifically, \code{prm} includes three elements named "name", "pos", and "val".
The first two elements determine the exact parameter that changes, while the
third one is a numeric vector holding the values it takes. For example,
in the case of the Multivariate Gamma distribution,
\code{D0 <- MGamma(shape = c(1, 2), scale = 3)} and
\code{prm <- list(name = "shape", pos = 2, val = seq(1, 1.5, by = 0.1))}
means that the evaluation will be performed for the MGamma distributions with
shape parameters \verb{(1, 1)}, \verb{(1, 1.1)}, ..., \verb{(1, 1.5)} and scale \code{3}. Notice
that the initial shape parameter \code{2} in \code{D0} is not utilized in the function.
}
\examples{
\dontrun{
# Distribution
d_beta <- Beta(shape1 = 1, shape2 = 1.5)

# Simulation
set.seed(2)
x <- r(d_beta)(50)

# Estimation
me(x, "Beta")
same(x, "Beta")
mle(x, "Beta")

# Asymptotic Covariance Matrix
acov_me(d_beta)
acov_same(d_beta)
acov_mle(d_beta)
}
}
\seealso{
\code{\link[=plot_metrics]{plot_metrics()}}
}
