% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle}
\alias{mle}
\alias{mle,ANY,character-method}
\alias{mle,numeric,Beta-method}
\alias{mle,matrix,Beta-method}
\alias{mle,numeric,Gammad-method}
\alias{mle,matrix,Gammad-method}
\alias{mle,matrix,Dirichlet-method}
\alias{mle,array,MGamma-method}
\title{Maximum Likelihood Estimator}
\usage{
mle(x, distr, ...)

\S4method{mle}{ANY,character}(x, distr, ...)

\S4method{mle}{numeric,Beta}(
  x,
  distr,
  par0 = same,
  method = "L-BFGS-B",
  lower = c(1e-05, 1e-05),
  upper = c(Inf, Inf)
)

\S4method{mle}{matrix,Beta}(
  x,
  distr,
  par0 = same,
  method = "L-BFGS-B",
  lower = c(1e-05, 1e-05),
  upper = c(Inf, Inf)
)

\S4method{mle}{numeric,Gammad}(
  x,
  distr,
  par0 = same,
  method = "L-BFGS-B",
  lower = c(1e-05, 1e-05),
  upper = c(Inf, Inf)
)

\S4method{mle}{matrix,Gammad}(
  x,
  distr,
  par0 = same,
  method = "L-BFGS-B",
  lower = c(1e-05, 1e-05),
  upper = c(Inf, Inf)
)

\S4method{mle}{matrix,Dirichlet}(
  x,
  distr,
  par0 = same,
  method = "L-BFGS-B",
  lower = rep(1e-05, times = nrow(x)),
  upper = rep(Inf, times = nrow(x))
)

\S4method{mle}{array,MGamma}(x, distr, par0 = same, method = "L-BFGS-B", lower = NULL, upper = NULL)
}
\arguments{
\item{x}{numeric. A sample under estimation.}

\item{distr}{A subclass of \code{Distribution}. The distribution of interest.}

\item{...}{extra arguments.}

\item{margin}{numeric. Used if \code{x} is a matrix. Passed to \code{apply}.}
}
\value{
numeric. The estimator produced by the sample.
}
\description{
Maximum Likelihood Estimator
}
