% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Multigam.R
\name{Multigam}
\alias{Multigam}
\alias{dmultigam}
\alias{rmultigam}
\alias{d,Multigam,numeric-method}
\alias{r,Multigam,numeric-method}
\alias{mean,Multigam-method}
\alias{var,Multigam-method}
\alias{finf,Multigam-method}
\alias{llmultigam}
\alias{ll,Multigam,matrix-method}
\alias{emultigam}
\alias{mle,Multigam,matrix-method}
\alias{me,Multigam,matrix-method}
\alias{same,Multigam,matrix-method}
\alias{vmultigam}
\alias{avar_mle,Multigam-method}
\alias{avar_me,Multigam-method}
\alias{avar_same,Multigam-method}
\title{Gamma Distribution}
\usage{
Multigam(shape = 1, scale = 1)

dmultigam(x, shape, scale, log = FALSE)

rmultigam(n, shape, scale)

\S4method{d}{Multigam,numeric}(distr, x)

\S4method{r}{Multigam,numeric}(distr, n)

\S4method{mean}{Multigam}(x)

\S4method{var}{Multigam}(x)

\S4method{finf}{Multigam}(x)

llmultigam(x, shape, scale)

\S4method{ll}{Multigam,matrix}(distr, x)

emultigam(x, type = "mle", ...)

\S4method{mle}{Multigam,matrix}(distr, x, par0 = "same", method = "L-BFGS-B", lower = 1e-05, upper = Inf)

\S4method{me}{Multigam,matrix}(distr, x)

\S4method{same}{Multigam,matrix}(distr, x)

vmultigam(shape, scale, type = "mle")

\S4method{avar_mle}{Multigam}(distr)

\S4method{avar_me}{Multigam}(distr)

\S4method{avar_same}{Multigam}(distr)
}
\arguments{
\item{shape, scale}{numeric. The distribution parameters.}

\item{x}{an object of class \code{Multigam}. If the function also has a \code{distr}
argument, \code{x} is a numeric vector, a sample of observations.}

\item{log}{logical. Should the log of the density be returned?}

\item{n}{numeric. The sample size.}

\item{distr}{an object of class \code{Multigam}.}

\item{type}{character, case ignored. The estimator type (mle, me, or same).}

\item{...}{extra arguments.}

\item{par0, method, lower, upper}{arguments passed to optim.}
}
\value{
Each type of function returns a different type of object:
\itemize{
\item Distribution Functions: When supplied with one argument (\code{distr}), the
\code{d()}, \code{p()}, \code{q()}, \code{r()}, \code{ll()} functions return the density, cumulative
probability, quantile, random sample generator, and log-likelihood functions,
respectively. When supplied with both arguments (\code{distr} and \code{x}), they
evaluate the aforementioned functions directly.
\item Moments: Returns a numeric, either vector or matrix depending on the moment
and the distribution. The \code{moments()} function returns a list with all the
available methods.
\item Estimation: Returns a list. The estimator of the unknown parameters. Note
that in distribution families like the binomial, multinomial, and negative
binomial, the size is not returned, since it is considered known.
\item Variance: Returns a named matrix. The asymptotic covariance matrix of the
estimator.
}
}
\description{
Gamma Distribution
}
