% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Nbinom.R
\name{Nbinom}
\alias{Nbinom}
\alias{d,Nbinom,numeric-method}
\alias{p,Nbinom,numeric-method}
\alias{qn,Nbinom,numeric-method}
\alias{r,Nbinom,numeric-method}
\alias{mean,Nbinom-method}
\alias{median,Nbinom-method}
\alias{mode,Nbinom-method}
\alias{var,Nbinom-method}
\alias{sd,Nbinom-method}
\alias{skew,Nbinom-method}
\alias{kurt,Nbinom-method}
\alias{entro,Nbinom-method}
\alias{finf,Nbinom-method}
\alias{llnbinom}
\alias{ll,Nbinom,numeric-method}
\alias{enbinom}
\alias{mle,Nbinom,numeric-method}
\alias{me,Nbinom,numeric-method}
\alias{vnbinom}
\alias{avar_mle,Nbinom-method}
\alias{avar_me,Nbinom-method}
\title{Negative Binomial Distribution}
\usage{
Nbinom(size = 1, prob = 0.5)

\S4method{d}{Nbinom,numeric}(distr, x)

\S4method{p}{Nbinom,numeric}(distr, x)

\S4method{qn}{Nbinom,numeric}(distr, x)

\S4method{r}{Nbinom,numeric}(distr, n)

\S4method{mean}{Nbinom}(x)

\S4method{median}{Nbinom}(x)

\S4method{mode}{Nbinom}(x)

\S4method{var}{Nbinom}(x)

\S4method{sd}{Nbinom}(x)

\S4method{skew}{Nbinom}(x)

\S4method{kurt}{Nbinom}(x)

\S4method{entro}{Nbinom}(x)

\S4method{finf}{Nbinom}(x)

llnbinom(x, size, prob)

\S4method{ll}{Nbinom,numeric}(distr, x)

enbinom(x, size, type = "mle", ...)

\S4method{mle}{Nbinom,numeric}(distr, x)

\S4method{me}{Nbinom,numeric}(distr, x)

vnbinom(size, prob, type = "mle")

\S4method{avar_mle}{Nbinom}(distr)

\S4method{avar_me}{Nbinom}(distr)
}
\arguments{
\item{size, prob}{numeric. The distribution parameters.}

\item{distr}{an object of class \code{Nbinom}.}

\item{x}{an object of class \code{Nbinom}. If the function also has a \code{distr}
argument, \code{x} is a numeric vector, a sample of observations.}

\item{n}{numeric. The sample size.}

\item{type}{character, case ignored. The estimator type (mle, me, or same).}

\item{...}{extra arguments.}
}
\value{
Each type of function returns a different type of object:
\itemize{
\item Distribution Functions: When supplied with one argument (\code{distr}), the
\code{d()}, \code{p()}, \code{q()}, \code{r()}, \code{ll()} functions return the density, cumulative
probability, quantile, random sample generator, and log-likelihood functions,
respectively. When supplied with both arguments (\code{distr} and \code{x}), they
evaluate the aforementioned functions directly.
\item Moments: Returns a numeric, either vector or matrix depending on the moment
and the distribution. The \code{moments()} function returns a list with all the
available methods.
\item Estimation: Returns a list. The estimator of the unknown parameters. Note
that in distribution families like the binomial, multinomial, and negative
binomial, the size is not returned, since it is considered known.
\item Variance: Returns a named matrix. The asymptotic covariance matrix of the
estimator.
}
}
\description{
Negative Binomial Distribution
}
