% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acov.R
\name{plot_acov}
\alias{plot_acov}
\title{Plot the Asymptotic Covariance}
\usage{
plot_acov(
  x,
  colors,
  save = FALSE,
  path = getwd(),
  name = "myplot.pdf",
  width = 15,
  height = 8
)
}
\arguments{
\item{x}{A data.frame. The result of \code{acov()}.}

\item{colors}{character. The colors to be used in the plot.}

\item{save}{logical. Should the plot be saved?}

\item{path}{A path to the directory in which the plot will be saved.}

\item{name}{character. The name of the output pdf file.}

\item{width}{numeric. The plot width in inches.}

\item{height}{numeric. The plot height in inches.}
}
\value{
The plot is returned invisibly in the form of a \code{ggplot} object.
}
\description{
This function provides an easy way to illustrate the output of \code{acov()},
using the \code{ggplot2} package. A line chart is created in which each estimator
is plotted with a different color and line type. The plot can be saved in
pdf format.
}
\examples{
\dontrun{
# Distribution
D <- Beta(shape1 = 1, shape2 = 1.5)

# Simulation
set.seed(2)
x <- r(D)(50)

# Estimation
me(x, "Beta")
same(x, "Beta")
mle(x, "Beta")

# Asymptotic Covariance Matrix
acov_me(D)
acov_same(D)
acov_mle(D)
}
}
\seealso{
\code{\link[=acov]{acov()}}
}
