% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Pois.R
\name{Pois}
\alias{Pois}
\alias{d,Pois,numeric-method}
\alias{p,Pois,numeric-method}
\alias{qn,Pois,numeric-method}
\alias{r,Pois,numeric-method}
\alias{mean,Pois-method}
\alias{median,Pois-method}
\alias{mode,Pois-method}
\alias{var,Pois-method}
\alias{sd,Pois-method}
\alias{skew,Pois-method}
\alias{kurt,Pois-method}
\alias{entro,Pois-method}
\alias{finf,Pois-method}
\alias{llpois}
\alias{ll,Pois,numeric-method}
\alias{epois}
\alias{mle,Pois,numeric-method}
\alias{me,Pois,numeric-method}
\alias{vpois}
\alias{avar_mle,Pois-method}
\alias{avar_me,Pois-method}
\title{Poisson Distribution}
\usage{
Pois(lambda = 1)

\S4method{d}{Pois,numeric}(distr, x)

\S4method{p}{Pois,numeric}(distr, x)

\S4method{qn}{Pois,numeric}(distr, x)

\S4method{r}{Pois,numeric}(distr, n)

\S4method{mean}{Pois}(x)

\S4method{median}{Pois}(x)

\S4method{mode}{Pois}(x)

\S4method{var}{Pois}(x)

\S4method{sd}{Pois}(x)

\S4method{skew}{Pois}(x)

\S4method{kurt}{Pois}(x)

\S4method{entro}{Pois}(x)

\S4method{finf}{Pois}(x)

llpois(x, lambda)

\S4method{ll}{Pois,numeric}(distr, x)

epois(x, type = "mle", ...)

\S4method{mle}{Pois,numeric}(distr, x)

\S4method{me}{Pois,numeric}(distr, x)

vpois(lambda, type = "mle")

\S4method{avar_mle}{Pois}(distr)

\S4method{avar_me}{Pois}(distr)
}
\arguments{
\item{lambda}{numeric. The distribution parameters.}

\item{distr}{an object of class \code{Pois}.}

\item{x}{an object of class \code{Pois}. If the function also has a \code{distr}
argument, \code{x} is a numeric vector, a sample of observations.}

\item{n}{numeric. The sample size.}

\item{type}{character, case ignored. The estimator type (mle, me, or same).}

\item{...}{extra arguments.}
}
\value{
Each type of function returns a different type of object:
\itemize{
\item Distribution Functions: When supplied with one argument (\code{distr}), the
\code{d()}, \code{p()}, \code{q()}, \code{r()}, \code{ll()} functions return the density, cumulative
probability, quantile, random sample generator, and log-likelihood functions,
respectively. When supplied with both arguments (\code{distr} and \code{x}), they
evaluate the aforementioned functions directly.
\item Moments: Returns a numeric, either vector or matrix depending on the moment
and the distribution. The \code{moments()} function returns a list with all the
available methods.
\item Estimation: Returns a list. The estimator of the unknown parameters. Note
that in distribution families like the binomial, multinomial, and negative
binomial, the size is not returned, since it is considered known.
\item Variance: Returns a named matrix. The asymptotic covariance matrix of the
estimator.
}
}
\description{
Poisson Distribution
}
