% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_generics.R, R/02_Beta.R, R/02_Dirichlet.R,
%   R/02_Gammad.R, R/02_MGamma.R
\name{same}
\alias{same}
\alias{same,ANY,character-method}
\alias{same,numeric,Beta-method}
\alias{same,matrix,Dirichlet-method}
\alias{same,numeric,Gammad-method}
\alias{same,matrix,MGamma-method}
\title{Score - Adjusted Moment Estimation}
\usage{
same(x, distr, ...)

\S4method{same}{ANY,character}(x, distr, ...)

\S4method{same}{numeric,Beta}(x, distr)

\S4method{same}{matrix,Dirichlet}(x, distr)

\S4method{same}{numeric,Gammad}(x, distr)

\S4method{same}{matrix,MGamma}(x, distr, dirich = FALSE)
}
\arguments{
\item{x}{numeric. A sample under estimation.}

\item{distr}{A subclass of \code{Distribution}. The distribution family assumed.}

\item{...}{extra arguments.}

\item{dirich}{logical. Should the Dirichlet-based estimator be calculated
instead? Applies only to the Multivariate Gamma distribution.}
}
\value{
numeric. The estimator produced by the sample.
}
\description{
Calculates the SAME under the assumption the sample observations are
independent and identically distributed (iid) according to a
specified family of distributions.
}
\examples{
\dontrun{
# -------------------------------------------
# Beta Distribution Example
# -------------------------------------------

# Simulation

set.seed(1)
x <- rbeta(100, shape1, shape2)
D <- Beta(shape1, shape2)

# Likelihood - The ll Functions

llbeta(x, shape1, shape2)
ll(x, c(shape1, shape2), D)
ll(x, c(shape1, shape2), "beta")

# Point Estimation - The e Functions

ebeta(x, type = "mle")
ebeta(x, type = "me")
ebeta(x, type = "same")

mle(x, D)
me(x, D)
same(x, D)

estim(x, D, type = "mle")

# Asymptotic Variance - The v Functions

vbeta(shape1, shape2, type = "mle")
vbeta(shape1, shape2, type = "me")
vbeta(shape1, shape2, type = "same")

avar_mle(D)
avar_me(D)
avar_same(D)

avar(D, type = "mle")
}
}
\seealso{
\link{estim}, \link{mle}, \link{me}
}
