% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Stud.R
\name{Stud}
\alias{Stud}
\alias{d,Stud,numeric-method}
\alias{p,Stud,numeric-method}
\alias{qn,Stud,numeric-method}
\alias{r,Stud,numeric-method}
\alias{mean,Stud-method}
\alias{median,Stud-method}
\alias{mode,Stud-method}
\alias{var,Stud-method}
\alias{sd,Stud-method}
\alias{skew,Stud-method}
\alias{kurt,Stud-method}
\alias{entro,Stud-method}
\alias{llt}
\alias{ll,Stud,numeric-method}
\title{Student Distribution}
\usage{
Stud(df = 1)

\S4method{d}{Stud,numeric}(distr, x)

\S4method{p}{Stud,numeric}(distr, x)

\S4method{qn}{Stud,numeric}(distr, x)

\S4method{r}{Stud,numeric}(distr, n)

\S4method{mean}{Stud}(x)

\S4method{median}{Stud}(x)

\S4method{mode}{Stud}(x)

\S4method{var}{Stud}(x)

\S4method{sd}{Stud}(x)

\S4method{skew}{Stud}(x)

\S4method{kurt}{Stud}(x)

\S4method{entro}{Stud}(x)

llt(x, df)

\S4method{ll}{Stud,numeric}(distr, x)
}
\arguments{
\item{df}{numeric. The distribution parameter.}

\item{distr}{an object of class \code{Stud}.}

\item{x}{an object of class \code{Stud}. If the function also has a \code{distr}
argument, \code{x} is a numeric vector, a sample of observations.}

\item{n}{numeric. The sample size.}
}
\value{
Each type of function returns a different type of object:
\itemize{
\item Distribution Functions: When supplied with one argument (\code{distr}), the
\code{d()}, \code{p()}, \code{q()}, \code{r()}, \code{ll()} functions return the density, cumulative
probability, quantile, random sample generator, and log-likelihood functions,
respectively. When supplied with both arguments (\code{distr} and \code{x}), they
evaluate the aforementioned functions directly.
\item Moments: Returns a numeric, either vector or matrix depending on the moment
and the distribution. The \code{moments()} function returns a list with all the
available methods.
\item Estimation: Returns a list. The estimator of the unknown parameters. Note
that in distribution families like the binomial, multinomial, and negative
binomial, the size is not returned, since it is considered known.
\item Variance: Returns a named matrix. The asymptotic covariance matrix of the
estimator.
}
}
\description{
Student Distribution
}
