% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Weib.R
\name{Weib}
\alias{Weib}
\alias{d,Weib-method}
\alias{p,Weib-method}
\alias{q2,Weib-method}
\alias{r,Weib-method}
\alias{mean,Weib-method}
\alias{median,Weib-method}
\alias{mode,Weib-method}
\alias{var,Weib-method}
\alias{sd,Weib-method}
\alias{skew,Weib-method}
\alias{kurt,Weib-method}
\alias{entro,Weib-method}
\alias{ll,numeric,numeric,Weib-method}
\alias{mle,numeric,Weib-method}
\alias{me,numeric,Weib-method}
\alias{avar_mle,Weib-method}
\alias{avar_me,Weib-method}
\title{Weibull Distribution}
\usage{
Weib(shape = 1, scale = 1)

\S4method{d}{Weib}(x)

\S4method{p}{Weib}(x)

\S4method{q2}{Weib}(x)

\S4method{r}{Weib}(x)

\S4method{mean}{Weib}(x)

\S4method{median}{Weib}(x)

\S4method{mode}{Weib}(x)

\S4method{var}{Weib}(x)

\S4method{sd}{Weib}(x)

\S4method{skew}{Weib}(x)

\S4method{kurt}{Weib}(x)

\S4method{entro}{Weib}(x)

\S4method{ll}{numeric,numeric,Weib}(x, prm, distr)

\S4method{mle}{numeric,Weib}(x, distr, par0 = "same", method = "L-BFGS-B", lower = 1e-05, upper = Inf)

\S4method{me}{numeric,Weib}(x, distr)

\S4method{avar_mle}{Weib}(distr)

\S4method{avar_me}{Weib}(distr)
}
\arguments{
\item{shape, scale}{numeric. The distribution parameters.}

\item{x}{an object of class \code{Weib}. If the function also has a \code{distr}
argument, \code{x} is a numeric vector, a sample of observations.}

\item{prm}{numeric. A vector including the distribution parameters.}

\item{distr}{an object of class \code{Weib}.}

\item{par0, method, lower, upper}{arguments passed to optim.}
}
\description{
Weibull Distribution
}
