% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.R
\name{acov_me}
\alias{acov_me}
\alias{acov_me,Beta-method}
\alias{acov_me,Gammad-method}
\alias{acov_me,Dirichlet-method}
\alias{acov_me,MGamma-method}
\title{Asymptotic covariance matrix of estimator}
\usage{
acov_me(distr, ...)

\S4method{acov_me}{Beta}(distr)

\S4method{acov_me}{Gammad}(distr)

\S4method{acov_me}{Dirichlet}(distr)

\S4method{acov_me}{MGamma}(distr)
}
\arguments{
\item{distr}{A subclass of \code{Distribution}. The distribution of interest.}

\item{...}{extra arguments.}
}
\value{
matrix. The asymptotic covariance matrix of the estimator.
}
\description{
Asymptotic covariance matrix of estimator
}
