% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_generics.R, R/02_Beta.R, R/02_Dirichlet.R,
%   R/02_Gamma.R, R/02_MGamma.R
\name{acov_mle}
\alias{acov_mle}
\alias{acov_mle,Beta-method}
\alias{acov_mle,Dirichlet-method}
\alias{acov_mle,Gammad-method}
\alias{acov_mle,MGamma-method}
\title{Asymptotic covariance matrix of estimator}
\usage{
acov_mle(distr, ...)

\S4method{acov_mle}{Beta}(distr)

\S4method{acov_mle}{Dirichlet}(distr)

\S4method{acov_mle}{Gammad}(distr)

\S4method{acov_mle}{MGamma}(distr)
}
\arguments{
\item{distr}{A subclass of \code{Distribution}. The distribution of interest.}

\item{...}{extra arguments.}
}
\value{
A named matrix. The asymptotic covariance matrix of the estimator.
}
\description{
Asymptotic covariance matrix of estimator
}
\examples{
\dontrun{
# Distribution
D <- Beta(shape1 = 1, shape2 = 1.5)

# Simulation
set.seed(2)
x <- r(D)(50)

# Estimation
me(x, "Beta")
same(x, "Beta")
mle(x, "Beta")

# Asymptotic Covariance Matrix
acov_me(D)
acov_same(D)
acov_mle(D)
}
}
