% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_generics.R
\name{avar}
\alias{avar}
\alias{avar_mle}
\alias{avar_me}
\alias{avar_same}
\title{Asymptotic Variance}
\usage{
avar(distr, type, ...)

avar_mle(distr, ...)

avar_me(distr, ...)

avar_same(distr, ...)
}
\arguments{
\item{distr}{A \code{Distribution} object.}

\item{type}{character, case ignored. The estimator type (\code{"mle"}, \code{"me"}, or \code{"same"}).}

\item{...}{extra arguments.}
}
\value{
A named matrix. The asymptotic covariance matrix of the estimator.
}
\description{
These functions calculate the asymptotic variance (or variance - covariance
matrix in the multidimensional case) of an estimator, given a specified
family of distributions and the true parameter values.
}
\section{Functions}{
\itemize{
\item \code{avar_mle()}: Asymptotic Variance of the Maximum Likelihood Estimator

\item \code{avar_me()}: Asymptotic Variance of the Moment Estimator

\item \code{avar_same()}: Asymptotic Variance of the Score-Adjusted Moment Estimator

}}
\examples{
# -----------------------------------------------------
# Beta Distribution Example
# -----------------------------------------------------

# Simulation
set.seed(1)
a <- 1
b <- 2
D <- Beta(a, b)
x <- r(D)(100)

# Point Estimation - The e Functions

ebeta(x, type = "mle")
ebeta(x, type = "me")
ebeta(x, type = "same")

mle(D, x)
me(D, x)
same(D, x)
e(D, x, type = "mle")

mle("beta", x) # the distr argument can be a character

# Asymptotic Variance - The v Functions

vbeta(a, b, type = "mle")
vbeta(a, b, type = "me")
vbeta(a, b, type = "same")

avar_mle(D)
avar_me(D)
avar_same(D)

avar(D, type = "mle")
}
\references{
General Textbooks
\itemize{
\item Van der Vaart, A. W. (2000), Asymptotic statistics, Vol. 3,
Cambridge university press.
}

Beta and gamma distribution families
\itemize{
\item Ye, Z.-S. & Chen, N. (2017), Closed-form estimators for the gamma
distribution derived from likelihood equations, The American Statistician
71(2), 177–181.
\item Tamae, H., Irie, K. & Kubokawa, T. (2020), A score-adjusted approach to
closed-form estimators for the gamma and beta distributions, Japanese Journal
of Statistics and Data Science 3, 543–561.
\item Mathal, A. & Moschopoulos, P. (1992), A form of multivariate gamma
distribution, Annals of the Institute of Statistical Mathematics 44, 97–106.
\item Oikonomidis, I. & Trevezas, S. (2023), Moment-Type Estimators for the
Dirichlet and the Multivariate Gamma Distributions, arXiv,
https://arxiv.org/abs/2311.15025
}
}
\seealso{
\link{avar_mle}, \link{avar_me}, \link{avar_same}
}
