% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_generics.R, R/02_Beta.R, R/02_Binom.R,
%   R/02_Dirichlet.R, R/02_Exp.R, R/02_Gammad.R, R/02_MGamma.R, R/02_Norm.R,
%   R/02_Pois.R
\name{avar_me}
\alias{avar_me}
\alias{avar_me,Beta-method}
\alias{avar_me,Binom-method}
\alias{avar_me,Dirichlet-method}
\alias{avar_me,Exp-method}
\alias{avar_me,Gammad-method}
\alias{avar_me,MGamma-method}
\alias{avar_me,Norm-method}
\alias{avar_me,Pois-method}
\title{ME Asymptotic Variance}
\usage{
avar_me(distr, ...)

\S4method{avar_me}{Beta}(distr)

\S4method{avar_me}{Binom}(distr)

\S4method{avar_me}{Dirichlet}(distr, comp = FALSE)

\S4method{avar_me}{Exp}(distr)

\S4method{avar_me}{Gammad}(distr)

\S4method{avar_me}{MGamma}(distr, dirich = FALSE)

\S4method{avar_me}{Norm}(distr)

\S4method{avar_me}{Pois}(distr)
}
\arguments{
\item{distr}{A subclass of \code{Distribution}. The distribution family assumed.}

\item{...}{extra arguments.}

\item{comp}{logical. Should the component matrices A, B be returned instead?
This argument is used internally by the \code{avar_me} method for the
Multivariate Gamma Dirichlet-based ME that needs these matrices.}

\item{dirich}{logical. Should the Dirichlet-based estimator be calculated
instead? Applies only to the Multivariate Gamma distribution.}
}
\value{
A named matrix. The asymptotic covariance matrix of the estimator.
}
\description{
Calculates the asymptotic variance (or variance - covariance matrix in the
multidimensional case) of the ME, given a specified family of
distributions and the true parameter values.
}
\examples{
\dontrun{
# -------------------------------------------
# Beta Distribution Example
# -------------------------------------------

# Simulation

set.seed(1)
x <- rbeta(100, shape1, shape2)
D <- Beta(shape1, shape2)

# Likelihood - The ll Functions

llbeta(x, shape1, shape2)
ll(x, c(shape1, shape2), D)
ll(x, c(shape1, shape2), "beta")

# Point Estimation - The e Functions

ebeta(x, type = "mle")
ebeta(x, type = "me")
ebeta(x, type = "same")

mle(x, D)
me(x, D)
same(x, D)

estim(x, D, type = "mle")

# Asymptotic Variance - The v Functions

vbeta(shape1, shape2, type = "mle")
vbeta(shape1, shape2, type = "me")
vbeta(shape1, shape2, type = "same")

avar_mle(D)
avar_me(D)
avar_same(D)

avar(D, type = "mle")
}
}
\seealso{
\link{avar}, \link{avar_mle}, \link{avar_same}
}
