% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_generics.R, R/02_Beta.R, R/02_Binom.R,
%   R/02_Dirichlet.R, R/02_Exp.R, R/02_Gammad.R, R/02_MGamma.R, R/02_Norm.R,
%   R/02_Pois.R
\name{avar_mle}
\alias{avar_mle}
\alias{avar_mle,Beta-method}
\alias{avar_mle,Binom-method}
\alias{avar_mle,Dirichlet-method}
\alias{avar_mle,Exp-method}
\alias{avar_mle,Gammad-method}
\alias{avar_mle,MGamma-method}
\alias{avar_mle,Norm-method}
\alias{avar_mle,Pois-method}
\title{MLE Asymptotic Variance}
\usage{
avar_mle(distr, ...)

\S4method{avar_mle}{Beta}(distr)

\S4method{avar_mle}{Binom}(distr)

\S4method{avar_mle}{Dirichlet}(distr)

\S4method{avar_mle}{Exp}(distr)

\S4method{avar_mle}{Gammad}(distr)

\S4method{avar_mle}{MGamma}(distr)

\S4method{avar_mle}{Norm}(distr)

\S4method{avar_mle}{Pois}(distr)
}
\arguments{
\item{distr}{A subclass of \code{Distribution}. The distribution family assumed.}

\item{...}{extra arguments.}
}
\value{
A named matrix. The asymptotic covariance matrix of the estimator.
}
\description{
Calculates the asymptotic variance (or variance - covariance matrix in the
multidimensional case) of the MLE, given a specified family of
distributions and the true parameter values.
}
\examples{
\dontrun{
# -------------------------------------------
# Beta Distribution Example
# -------------------------------------------

# Simulation

set.seed(1)
x <- rbeta(100, shape1, shape2)
D <- Beta(shape1, shape2)

# Likelihood - The ll Functions

llbeta(x, shape1, shape2)
ll(x, c(shape1, shape2), D)
ll(x, c(shape1, shape2), "beta")

# Point Estimation - The e Functions

ebeta(x, type = "mle")
ebeta(x, type = "me")
ebeta(x, type = "same")

mle(x, D)
me(x, D)
same(x, D)

estim(x, D, type = "mle")

# Asymptotic Variance - The v Functions

vbeta(shape1, shape2, type = "mle")
vbeta(shape1, shape2, type = "me")
vbeta(shape1, shape2, type = "same")

avar_mle(D)
avar_me(D)
avar_same(D)

avar(D, type = "mle")
}
}
\seealso{
\link{avar}, \link{avar_me}, \link{avar_same}
}
