% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Bern.R
\name{Bern}
\alias{Bern}
\alias{d,Bern,numeric-method}
\alias{p,Bern,numeric-method}
\alias{qn,Bern,numeric-method}
\alias{r,Bern,numeric-method}
\alias{mean,Bern-method}
\alias{median,Bern-method}
\alias{mode,Bern-method}
\alias{var,Bern-method}
\alias{sd,Bern-method}
\alias{skew,Bern-method}
\alias{kurt,Bern-method}
\alias{entro,Bern-method}
\alias{finf,Bern-method}
\alias{llbern}
\alias{ll,Bern,numeric-method}
\alias{ebern}
\alias{mle,Bern,numeric-method}
\alias{me,Bern,numeric-method}
\alias{vbern}
\alias{avar_mle,Bern-method}
\alias{avar_me,Bern-method}
\title{Bernoulli Distribution}
\usage{
Bern(prob = 0.5)

\S4method{d}{Bern,numeric}(distr, x)

\S4method{p}{Bern,numeric}(distr, x)

\S4method{qn}{Bern,numeric}(distr, x)

\S4method{r}{Bern,numeric}(distr, n)

\S4method{mean}{Bern}(x)

\S4method{median}{Bern}(x)

\S4method{mode}{Bern}(x)

\S4method{var}{Bern}(x)

\S4method{sd}{Bern}(x)

\S4method{skew}{Bern}(x)

\S4method{kurt}{Bern}(x)

\S4method{entro}{Bern}(x)

\S4method{finf}{Bern}(x)

llbern(x, prob)

\S4method{ll}{Bern,numeric}(distr, x)

ebern(x, type = "mle", ...)

\S4method{mle}{Bern,numeric}(distr, x)

\S4method{me}{Bern,numeric}(distr, x)

vbern(prob, type = "mle")

\S4method{avar_mle}{Bern}(distr)

\S4method{avar_me}{Bern}(distr)
}
\arguments{
\item{prob}{numeric. The distribution parameter.}

\item{distr}{an object of class \code{Bern}.}

\item{x}{an object of class \code{Bern}. If the function also has a \code{distr}
argument, \code{x} is a numeric vector, a sample of observations.}

\item{n}{numeric. The sample size.}

\item{type}{character, case ignored. The estimator type (mle, me, or same).}

\item{...}{extra arguments.}
}
\value{
Each type of function returns a different type of object:
\itemize{
\item Distribution Functions: When supplied with one argument (\code{distr}), the
\code{d()}, \code{p()}, \code{q()}, \code{r()}, \code{ll()} functions return the density, cumulative
probability, quantile, random sample generator, and log-likelihood functions,
respectively. When supplied with both arguments (\code{distr} and \code{x}), they
evaluate the aforementioned functions directly.
\item Moments: Returns a numeric, either vector or matrix depending on the moment
and the distribution. The \code{moments()} function returns a list with all the
available methods.
\item Estimation: Returns a list. The estimator of the unknown parameters. Note
that in distribution families like the binomial, multinomial, and negative
binomial, the size is not returned, since it is considered known.
\item Variance: Returns a named matrix. The asymptotic covariance matrix of the
estimator.
}
}
\description{
Bernoulli Distribution
}
\seealso{
\link{Distributions}, \link{moments}
}
