% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Cat.R
\name{Cat}
\alias{Cat}
\alias{d,Cat-method}
\alias{r,Cat-method}
\alias{mean,Cat-method}
\alias{var,Cat-method}
\alias{finf,Cat-method}
\alias{ll,numeric,numeric,Cat-method}
\alias{mle,numeric,Cat-method}
\alias{me,numeric,Cat-method}
\alias{avar_mle,Cat-method}
\alias{avar_me,Cat-method}
\title{Categorical Distribution}
\usage{
Cat(prob = c(0.5, 0.5))

\S4method{d}{Cat}(x)

\S4method{r}{Cat}(x)

\S4method{mean}{Cat}(x)

\S4method{var}{Cat}(x)

\S4method{finf}{Cat}(x)

\S4method{ll}{numeric,numeric,Cat}(x, prm, distr)

\S4method{mle}{numeric,Cat}(x, distr)

\S4method{me}{numeric,Cat}(x, distr)

\S4method{avar_mle}{Cat}(distr)

\S4method{avar_me}{Cat}(distr)
}
\arguments{
\item{prob}{numeric. The distribution parameters.}

\item{x}{an object of class \code{Cat}. If the function also has a \code{distr}
argument, \code{x} is a numeric vector, a sample of observations.}

\item{prm}{numeric. A vector including the distribution parameters.}

\item{distr}{an object of class \code{Cat}.}
}
\description{
Categorical Distribution
}
