% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Cat.R
\name{Cat}
\alias{Cat}
\alias{d,Cat,numeric-method}
\alias{r,Cat,numeric-method}
\alias{mean,Cat-method}
\alias{mode,Cat-method}
\alias{var,Cat-method}
\alias{entro,Cat-method}
\alias{finf,Cat-method}
\alias{llcat}
\alias{ll,Cat,numeric-method}
\alias{ecat}
\alias{mle,Cat,numeric-method}
\alias{me,Cat,numeric-method}
\alias{vcat}
\alias{avar_mle,Cat-method}
\alias{avar_me,Cat-method}
\title{Categorical Distribution}
\usage{
Cat(prob = c(0.5, 0.5))

\S4method{d}{Cat,numeric}(distr, x)

\S4method{r}{Cat,numeric}(distr, n)

\S4method{mean}{Cat}(x)

\S4method{mode}{Cat}(x)

\S4method{var}{Cat}(x)

\S4method{entro}{Cat}(x)

\S4method{finf}{Cat}(x)

llcat(x, prob)

\S4method{ll}{Cat,numeric}(distr, x)

ecat(x, type = "mle", ...)

\S4method{mle}{Cat,numeric}(distr, x, dim = NULL)

\S4method{me}{Cat,numeric}(distr, x, dim = NULL)

vcat(prob, type = "mle")

\S4method{avar_mle}{Cat}(distr)

\S4method{avar_me}{Cat}(distr)
}
\arguments{
\item{prob}{numeric. The distribution parameters.}

\item{distr}{an object of class \code{Cat}.}

\item{x}{an object of class \code{Cat}. If the function also has a \code{distr}
argument, \code{x} is a numeric vector, a sample of observations.}

\item{n}{numeric. The sample size.}

\item{type}{character, case ignored. The estimator type (mle, me, or same).}

\item{...}{extra arguments.}

\item{dim}{numeric. The parameter dimension. See details.}
}
\value{
Each type of function returns a different type of object:
\itemize{
\item Distribution Functions: When supplied with one argument (\code{distr}), the
\code{d()}, \code{p()}, \code{q()}, \code{r()}, \code{ll()} functions return the density, cumulative
probability, quantile, random sample generator, and log-likelihood functions,
respectively. When supplied with both arguments (\code{distr} and \code{x}), they
evaluate the aforementioned functions directly.
\item Moments: Returns a numeric, either vector or matrix depending on the moment
and the distribution. The \code{moments()} function returns a list with all the
available methods.
\item Estimation: Returns a list. The estimator of the unknown parameters. Note
that in distribution families like the binomial, multinomial, and negative
binomial, the size is not returned, since it is considered known.
\item Variance: Returns a named matrix. The asymptotic covariance matrix of the
estimator.
}
}
\description{
Categorical Distribution
}
\details{
The estimation of \code{prob} from a sample would by default return a vector of
probabilities corresponding to the categories that appeared in the sample and
0 for the rest. However, the parameter dimension cannot be uncovered by the
sample, it has to be provided separately. This can be done with the argument
\code{dim}. If \code{dim} is not supplied, the dimension will be retrieved from the
\code{distr} argument. Categories that did not appear in the sample will have 0
probabilities appended to the end of the prob vector.

Note that the actual dimension of the probability parameter vector is \code{k-1},
therefore the Fisher information matrix and the asymptotic variance -
covariance matrix of the estimators is of dimension \verb{(k-1)x(k-1)}.
}
