% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Cauchy.R
\name{Cauchy}
\alias{Cauchy}
\alias{d,Cauchy-method}
\alias{p,Cauchy-method}
\alias{q2,Cauchy-method}
\alias{r,Cauchy-method}
\alias{median,Cauchy-method}
\alias{mode,Cauchy-method}
\alias{entro,Cauchy-method}
\alias{finf,Cauchy-method}
\alias{ll,numeric,numeric,Cauchy-method}
\title{Cauchy Distribution}
\usage{
Cauchy(location = 1, scale = 1)

\S4method{d}{Cauchy}(x)

\S4method{p}{Cauchy}(x)

\S4method{q2}{Cauchy}(x)

\S4method{r}{Cauchy}(x)

\S4method{median}{Cauchy}(x)

\S4method{mode}{Cauchy}(x)

\S4method{entro}{Cauchy}(x)

\S4method{finf}{Cauchy}(x)

\S4method{ll}{numeric,numeric,Cauchy}(x, prm, distr)
}
\arguments{
\item{location, scale}{numeric. The distribution parameters.}

\item{x}{an object of class \code{Cauchy}. If the function also has a \code{distr}
argument, \code{x} is a numeric vector, a sample of observations.}

\item{prm}{numeric. A vector including the distribution parameters.}

\item{distr}{an object of class \code{Cauchy}.}
}
\description{
Cauchy Distribution
}
