% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Chisq.R
\name{Chisq}
\alias{Chisq}
\alias{d,Chisq,numeric-method}
\alias{p,Chisq,numeric-method}
\alias{qn,Chisq,numeric-method}
\alias{r,Chisq,numeric-method}
\alias{mean,Chisq-method}
\alias{median,Chisq-method}
\alias{mode,Chisq-method}
\alias{var,Chisq-method}
\alias{sd,Chisq-method}
\alias{skew,Chisq-method}
\alias{kurt,Chisq-method}
\alias{entro,Chisq-method}
\alias{finf,Chisq-method}
\alias{llchisq}
\alias{ll,Chisq,numeric-method}
\alias{echisq}
\alias{mle,Chisq,numeric-method}
\alias{me,Chisq,numeric-method}
\alias{vchisq}
\alias{avar_mle,Chisq-method}
\alias{avar_me,Chisq-method}
\title{Chi-Square Distribution}
\usage{
Chisq(df = 1)

\S4method{d}{Chisq,numeric}(distr, x)

\S4method{p}{Chisq,numeric}(distr, x)

\S4method{qn}{Chisq,numeric}(distr, x)

\S4method{r}{Chisq,numeric}(distr, n)

\S4method{mean}{Chisq}(x)

\S4method{median}{Chisq}(x)

\S4method{mode}{Chisq}(x)

\S4method{var}{Chisq}(x)

\S4method{sd}{Chisq}(x)

\S4method{skew}{Chisq}(x)

\S4method{kurt}{Chisq}(x)

\S4method{entro}{Chisq}(x)

\S4method{finf}{Chisq}(x)

llchisq(x, df)

\S4method{ll}{Chisq,numeric}(distr, x)

echisq(x, type = "mle", ...)

\S4method{mle}{Chisq,numeric}(distr, x)

\S4method{me}{Chisq,numeric}(distr, x)

vchisq(df, type = "mle")

\S4method{avar_mle}{Chisq}(distr)

\S4method{avar_me}{Chisq}(distr)
}
\arguments{
\item{df}{numeric. The distribution parameters.}

\item{distr}{an object of class \code{Chisq}.}

\item{x}{an object of class \code{Chisq}. If the function also has a \code{distr}
argument, \code{x} is a numeric vector, a sample of observations.}

\item{n}{numeric. The sample size.}

\item{type}{character, case ignored. The estimator type (mle, me, or same).}

\item{...}{extra arguments.}
}
\value{
Each type of function returns a different type of object:
\itemize{
\item Distribution Functions: When supplied with one argument (\code{distr}), the
\code{d()}, \code{p()}, \code{q()}, \code{r()}, \code{ll()} functions return the density, cumulative
probability, quantile, random sample generator, and log-likelihood functions,
respectively. When supplied with both arguments (\code{distr} and \code{x}), they
evaluate the aforementioned functions directly.
\item Moments: Returns a numeric, either vector or matrix depending on the moment
and the distribution. The \code{moments()} function returns a list with all the
available methods.
\item Estimation: Returns a list. The estimator of the unknown parameters. Note
that in distribution families like the binomial, multinomial, and negative
binomial, the size is not returned, since it is considered known.
\item Variance: Returns a named matrix. The asymptotic covariance matrix of the
estimator.
}
}
\description{
Chi-Square Distribution
}
