% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Chisq.R
\name{Chisq}
\alias{Chisq}
\alias{d,Chisq-method}
\alias{p,Chisq-method}
\alias{q2,Chisq-method}
\alias{r,Chisq-method}
\alias{mean,Chisq-method}
\alias{var,Chisq-method}
\alias{sd,Chisq-method}
\alias{skew,Chisq-method}
\alias{kurt,Chisq-method}
\title{Chi-Square Distribution}
\usage{
Chisq(df = 1, ncp = 0)

\S4method{d}{Chisq}(x)

\S4method{p}{Chisq}(x)

\S4method{q2}{Chisq}(x)

\S4method{r}{Chisq}(x)

\S4method{mean}{Chisq}(x)

\S4method{var}{Chisq}(x)

\S4method{sd}{Chisq}(x)

\S4method{skew}{Chisq}(x)

\S4method{kurt}{Chisq}(x)
}
\arguments{
\item{df, ncp}{numeric. The distribution parameters.}

\item{x}{an object of class \code{Chisq}. If the function also has a \code{distr}
argument, \code{x} is a numeric vector, a sample of observations.}
}
\description{
Chi-Square Distribution
}
