% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dMGamma}
\alias{dMGamma}
\alias{rMGamma}
\title{The Dirichlet Distribution}
\usage{
dMGamma(X, shape = 1, Sigma = diag(2), log = FALSE)

rMGamma(n, shape = 1, Sigma = diag(2))
}
\arguments{
\item{X}{matrix. The random matrix.}

\item{shape}{numeric. The shape parameter.}

\item{Sigma}{matrix. The matrix parameter.}

\item{log}{logical. If TRUE, probabilities p are given as log(p).}

\item{n}{numeric. The number of observations.}
}
\value{
\code{dMGamma} returns the evaluated density function.
\code{rMGamma} performs Monte-Carlo simulation.
}
\description{
The Dirichlet Distribution
}
\examples{
\dontrun{
# Classic R Stats Format
X <- matrix(c(2, 1, 3, 2), 2, 2)
dMGamma(X, shape = 3, Sigma = diag(2))
set.seed(1)
rMGamma(1, shape = 3, Sigma = diag(2))

# S4 Distribution Class
X <- matrix(c(2, 1, 3, 2), 2, 2)
d_mgamma <- MGamma(shape = 3, Sigma = diag(2))
d(d_mgamma)(X)
set.seed(1)
r(d_mgamma)(1)
}
}
