% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dMGamma}
\alias{dMGamma}
\alias{rMGamma}
\title{The Multivariate Gamma Distribution}
\usage{
dMGamma(x, shape, scale, log = FALSE)

rMGamma(n, shape, scale)
}
\arguments{
\item{x}{numeric. The quantile vector.}

\item{shape}{numeric. The shape parameter vector.}

\item{scale}{numeric. The scale parameter vector.}

\item{log}{logical. If \code{TRUE}, probabilities \code{p} are given as \code{log(p)}.}

\item{n}{numeric. The number of observations.}
}
\value{
\code{dMGamma} returns a numeric vector (the evaluated density function).
\code{rMGamma} returns a matrix with \code{length(shape)} rows and \code{n} columns.
}
\description{
Density function and random generation for the Multivariate Gamma
distribution with shape parameter vector \code{shape} and scale parameter \code{scale}.
}
\examples{
\dontrun{
# Classic R Stats Format
dMGamma(c(4, 6), shape = c(2, 3), scale = 2)
set.seed(1)
rMGamma(10, shape = c(2, 3), scale = 2)

# S4 Distribution Class
D <- MGamma(shape = c(2, 3), scale = 2)
d(D)(c(4, 6))
set.seed(1)
r(D)(10)
}
}
