% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators-package.R
\docType{package}
\name{estimators-package}
\alias{estimators}
\alias{estimators-package}
\title{estimators: Distribution Parameter Estimation Methods}
\description{
This package provides estimators for parameters of common distribution families.
}
\author{
\strong{Maintainer}: Ioannis Oikonomidis \email{goikon@math.uoa.gr} (\href{https://orcid.org/0000-0001-8130-2104}{ORCID})

Authors:
\itemize{
  \item Samis Trevezas \email{strevezas@math.uoa.gr} [thesis advisor]
}

}
\keyword{internal}
