% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Exp.R
\name{Exp}
\alias{Exp}
\alias{d,Exp-method}
\alias{p,Exp-method}
\alias{q2,Exp-method}
\alias{r,Exp-method}
\alias{mean,Exp-method}
\alias{median,Exp-method}
\alias{mode,Exp-method}
\alias{var,Exp-method}
\alias{sd,Exp-method}
\alias{skew,Exp-method}
\alias{kurt,Exp-method}
\alias{entro,Exp-method}
\alias{finf,Exp-method}
\alias{ll,numeric,numeric,Exp-method}
\alias{mle,numeric,Exp-method}
\alias{me,numeric,Exp-method}
\alias{avar_mle,Exp-method}
\alias{avar_me,Exp-method}
\title{Exponential Distribution}
\usage{
Exp(rate = 1)

\S4method{d}{Exp}(x)

\S4method{p}{Exp}(x)

\S4method{q2}{Exp}(x)

\S4method{r}{Exp}(x)

\S4method{mean}{Exp}(x)

\S4method{median}{Exp}(x)

\S4method{mode}{Exp}(x)

\S4method{var}{Exp}(x)

\S4method{sd}{Exp}(x)

\S4method{skew}{Exp}(x)

\S4method{kurt}{Exp}(x)

\S4method{entro}{Exp}(x)

\S4method{finf}{Exp}(x)

\S4method{ll}{numeric,numeric,Exp}(x, prm, distr)

\S4method{mle}{numeric,Exp}(x, distr)

\S4method{me}{numeric,Exp}(x, distr)

\S4method{avar_mle}{Exp}(distr)

\S4method{avar_me}{Exp}(distr)
}
\arguments{
\item{rate}{numeric. The distribution parameters.}

\item{x}{an object of class \code{Exp}. If the function also has a \code{distr}
argument, \code{x} is a numeric vector, a sample of observations.}

\item{prm}{numeric. A vector including the distribution parameters.}

\item{distr}{an object of class \code{Exp}.}
}
\description{
Exponential Distribution
}
