% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Fisher.R
\name{Fisher}
\alias{Fisher}
\alias{d,Fisher,numeric-method}
\alias{p,Fisher,numeric-method}
\alias{qn,Fisher,numeric-method}
\alias{r,Fisher,numeric-method}
\alias{mean,Fisher-method}
\alias{median,Fisher-method}
\alias{mode,Fisher-method}
\alias{var,Fisher-method}
\alias{sd,Fisher-method}
\alias{skew,Fisher-method}
\alias{kurt,Fisher-method}
\alias{entro,Fisher-method}
\alias{llf}
\alias{ll,Fisher,numeric-method}
\title{Fisher Distribution}
\usage{
Fisher(df1 = 1, df2 = 1)

\S4method{d}{Fisher,numeric}(distr, x)

\S4method{p}{Fisher,numeric}(distr, x)

\S4method{qn}{Fisher,numeric}(distr, x)

\S4method{r}{Fisher,numeric}(distr, n)

\S4method{mean}{Fisher}(x)

\S4method{median}{Fisher}(x)

\S4method{mode}{Fisher}(x)

\S4method{var}{Fisher}(x)

\S4method{sd}{Fisher}(x)

\S4method{skew}{Fisher}(x)

\S4method{kurt}{Fisher}(x)

\S4method{entro}{Fisher}(x)

llf(x, df1, df2)

\S4method{ll}{Fisher,numeric}(distr, x)
}
\arguments{
\item{df1, df2}{numeric. The distribution parameters.}

\item{distr}{an object of class \code{Fisher}.}

\item{x}{an object of class \code{Fisher}. If the function also has a \code{distr}
argument, \code{x} is a numeric vector, a sample of observations.}

\item{n}{numeric. The sample size.}
}
\value{
Each type of function returns a different type of object:
\itemize{
\item Distribution Functions: When supplied with one argument (\code{distr}), the
\code{d()}, \code{p()}, \code{q()}, \code{r()}, \code{ll()} functions return the density, cumulative
probability, quantile, random sample generator, and log-likelihood functions,
respectively. When supplied with both arguments (\code{distr} and \code{x}), they
evaluate the aforementioned functions directly.
\item Moments: Returns a numeric, either vector or matrix depending on the moment
and the distribution. The \code{moments()} function returns a list with all the
available methods.
\item Estimation: Returns a list. The estimator of the unknown parameters. Note
that in distribution families like the binomial, multinomial, and negative
binomial, the size is not returned, since it is considered known.
\item Variance: Returns a named matrix. The asymptotic covariance matrix of the
estimator.
}
}
\description{
Fisher Distribution
}
