% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Fisher.R
\name{Fisher}
\alias{Fisher}
\alias{d,Fisher-method}
\alias{p,Fisher-method}
\alias{q2,Fisher-method}
\alias{r,Fisher-method}
\alias{mean,Fisher-method}
\alias{mode,Fisher-method}
\alias{var,Fisher-method}
\alias{sd,Fisher-method}
\alias{skew,Fisher-method}
\title{Fisher Distribution}
\usage{
Fisher(df1 = 1, df2 = 1, ncp = 0)

\S4method{d}{Fisher}(x)

\S4method{p}{Fisher}(x)

\S4method{q2}{Fisher}(x)

\S4method{r}{Fisher}(x)

\S4method{mean}{Fisher}(x)

\S4method{mode}{Fisher}(x)

\S4method{var}{Fisher}(x)

\S4method{sd}{Fisher}(x)

\S4method{skew}{Fisher}(x)
}
\arguments{
\item{df1, df2, ncp}{numeric. The distribution parameters.}

\item{x}{an object of class \code{Fisher}. If the function also has a \code{distr}
argument, \code{x} is a numeric vector, a sample of observations.}
}
\description{
Fisher Distribution
}
