% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{ll}
\alias{ll}
\alias{ll,numeric,numeric,Beta-method}
\alias{ll,numeric,numeric,Gammad-method}
\alias{ll,numeric,matrix,Dirichlet-method}
\alias{ll,numeric,matrix,MGamma-method}
\title{Log-Likelihood}
\usage{
ll(prm, x, distr, ...)

\S4method{ll}{numeric,numeric,Beta}(prm, x, distr)

\S4method{ll}{numeric,numeric,Gammad}(prm, x, distr)

\S4method{ll}{numeric,matrix,Dirichlet}(prm, x, distr)

\S4method{ll}{numeric,matrix,MGamma}(prm, x, distr)
}
\arguments{
\item{prm}{numeric. A vector of the distribution parameters.}

\item{x}{numeric. A sample under estimation.}

\item{distr}{A subclass of \code{Distribution}. The distribution of interest.}

\item{...}{extra arguments.}
}
\value{
Numeric. The value of the log-likelihood function.
}
\description{
Log-Likelihood
}
