% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_generics.R, R/02_Beta.R, R/02_Dirichlet.R,
%   R/02_Gamma.R, R/02_MGamma.R
\name{me}
\alias{me}
\alias{me,ANY,character-method}
\alias{me2}
\alias{me2,ANY,character-method}
\alias{me,numeric,Beta-method}
\alias{me,matrix,Beta-method}
\alias{me,matrix,Dirichlet-method}
\alias{me,numeric,Gammad-method}
\alias{me,matrix,Gammad-method}
\alias{me,matrix,MGamma-method}
\alias{me2,matrix,MGamma-method}
\title{Moment Estimator}
\usage{
me(x, distr, ...)

\S4method{me}{ANY,character}(x, distr, ...)

me2(x, distr, ...)

\S4method{me2}{ANY,character}(x, distr, ...)

\S4method{me}{numeric,Beta}(x, distr)

\S4method{me}{matrix,Beta}(x, distr)

\S4method{me}{matrix,Dirichlet}(x, distr)

\S4method{me}{numeric,Gammad}(x, distr)

\S4method{me}{matrix,Gammad}(x, distr)

\S4method{me}{matrix,MGamma}(x, distr)

\S4method{me2}{matrix,MGamma}(x, distr)
}
\arguments{
\item{x}{numeric. A sample under estimation.}

\item{distr}{A subclass of \code{Distribution}. The distribution family assumed.}

\item{...}{extra arguments.}
}
\value{
numeric. The estimator produced by the sample.
}
\description{
Calculates the MΕ of a sample under the assumption the observations are
independent and identically distributed (iid) according to a specified
family of distributions.
}
\examples{
\dontrun{
# Distribution
D <- Beta(shape1 = 1, shape2 = 1.5)

# Simulation
set.seed(2)
x <- r(D)(50)

# Estimation
me(x, "Beta")
same(x, "Beta")
mle(x, "Beta")

# Asymptotic Covariance Matrix
acov_me(D)
acov_same(D)
acov_mle(D)
}
}
