% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\docType{class}
\name{MGamma-class}
\alias{MGamma-class}
\alias{MGamma}
\title{Matrix Gamma Distribution S4 Class}
\value{
An object of class \code{MGamma}.
}
\description{
Matrix Gamma Distribution S4 Class
}
\section{Slots}{

\describe{
\item{\code{shape}}{numeric. The shape parameter.}

\item{\code{Sigma}}{matrix. The matrix parameter.}
}}

\examples{
\dontrun{
# Classic R Stats Format
X <- matrix(c(2, 1, 3, 2), 2, 2)
dMGamma(X, shape = 3, Sigma = diag(2))
set.seed(1)
rMGamma(1, shape = 3, Sigma = diag(2))

# S4 Distribution Class
X <- matrix(c(2, 1, 3, 2), 2, 2)
d_mgamma <- MGamma(shape = 3, Sigma = diag(2))
d(d_mgamma)(X)
set.seed(1)
r(d_mgamma)(1)
}
}
