% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_generics.R, R/02_Beta.R, R/02_Dirichlet.R,
%   R/02_Gamma.R, R/02_MGamma.R
\name{mle}
\alias{mle}
\alias{mle,ANY,character-method}
\alias{mle,numeric,Beta-method}
\alias{mle,matrix,Beta-method}
\alias{mle,matrix,Dirichlet-method}
\alias{mle,numeric,Gammad-method}
\alias{mle,matrix,Gammad-method}
\alias{mle,matrix,MGamma-method}
\title{Maximum Likelihood Estimator}
\usage{
mle(x, distr, ...)

\S4method{mle}{ANY,character}(x, distr, ...)

\S4method{mle}{numeric,Beta}(
  x,
  distr,
  par0 = same,
  method = "L-BFGS-B",
  lower = c(1e-05, 1e-05),
  upper = c(Inf, Inf)
)

\S4method{mle}{matrix,Beta}(
  x,
  distr,
  par0 = same,
  method = "L-BFGS-B",
  lower = c(1e-05, 1e-05),
  upper = c(Inf, Inf)
)

\S4method{mle}{matrix,Dirichlet}(x, distr, par0 = "same", method = "L-BFGS-B", lower = 1e-05, upper = Inf)

\S4method{mle}{numeric,Gammad}(
  x,
  distr,
  par0 = same,
  method = "L-BFGS-B",
  lower = c(1e-05, 1e-05),
  upper = c(Inf, Inf)
)

\S4method{mle}{matrix,Gammad}(
  x,
  distr,
  par0 = same,
  method = "L-BFGS-B",
  lower = c(1e-05, 1e-05),
  upper = c(Inf, Inf)
)

\S4method{mle}{matrix,MGamma}(x, distr, par0 = "same", method = "L-BFGS-B", lower = 1e-05, upper = Inf)
}
\arguments{
\item{x}{numeric. A sample under estimation.}

\item{distr}{A subclass of \code{Distribution}. The distribution family assumed.}

\item{...}{extra arguments.}

\item{par0}{function. The estimator to use for initialization of the
likelihood maximization algorithm.}

\item{method, lower, upper}{arguments passed to optim.}
}
\value{
numeric. The estimator produced by the sample.
}
\description{
Calculates the MLE of a sample under the assumption the observations are
independent and identically distributed (iid) according to a specified
family of distributions.
}
\examples{
\dontrun{
# Distribution
D <- Beta(shape1 = 1, shape2 = 1.5)

# Simulation
set.seed(2)
x <- r(D)(50)

# Estimation
me(x, "Beta")
same(x, "Beta")
mle(x, "Beta")

# Asymptotic Covariance Matrix
acov_me(D)
acov_same(D)
acov_mle(D)
}
}
