% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Multigam.R
\name{Multigam}
\alias{Multigam}
\alias{dmultigam}
\alias{rmultigam}
\alias{d,Multigam-method}
\alias{r,Multigam-method}
\alias{mean,Multigam-method}
\alias{var,Multigam-method}
\alias{entro,Multigam-method}
\alias{finf,Multigam-method}
\alias{ll,matrix,numeric,Multigam-method}
\alias{mle,matrix,Multigam-method}
\alias{me,matrix,Multigam-method}
\alias{same,matrix,Multigam-method}
\alias{avar_mle,Multigam-method}
\alias{avar_me,Multigam-method}
\alias{avar_same,Multigam-method}
\title{Gamma Distribution}
\usage{
Multigam(shape = 1, scale = 1)

dmultigam(x, shape, scale, log = FALSE)

rmultigam(n, shape, scale)

\S4method{d}{Multigam}(x)

\S4method{r}{Multigam}(x)

\S4method{mean}{Multigam}(x)

\S4method{var}{Multigam}(x)

\S4method{entro}{Multigam}(x)

\S4method{finf}{Multigam}(x)

\S4method{ll}{matrix,numeric,Multigam}(x, prm, distr)

\S4method{mle}{matrix,Multigam}(x, distr, par0 = "same", method = "L-BFGS-B", lower = 1e-05, upper = Inf)

\S4method{me}{matrix,Multigam}(x, distr)

\S4method{same}{matrix,Multigam}(x, distr)

\S4method{avar_mle}{Multigam}(distr)

\S4method{avar_me}{Multigam}(distr)

\S4method{avar_same}{Multigam}(distr)
}
\arguments{
\item{shape, scale}{numeric. The distribution parameters.}

\item{x}{an object of class \code{Multigam}. If the function also has a \code{distr}
argument, \code{x} is a numeric vector, a sample of observations.}

\item{log}{logical. Should the log of the density be returned?}

\item{n}{numeric. The sample size.}

\item{prm}{numeric. A vector including the distribution parameters.}

\item{distr}{an object of class \code{Multigam}.}

\item{par0, method, lower, upper}{arguments passed to optim.}
}
\value{
The dpqr family of functions return the evaluated density, cumulative
probability, quantile, and random sample, respectively.
The moments family of functions return the appropriate theoretical moment,
as calculated by the distribution true parameters.
The ll function returns the evaluated log-likelihood, given a sample and the
theoretical parameters.
The estim family of functions return the estimated parameters of the
distribution, given a sample.
The avar family of functions return the asymptotic variance or variance -
covariance matrix (if there are two or more parameters) of the corresponding
estimation method.
Calculus performed on Distribution objects returns a Distribution object of
the appropriate class and with the appropriate parameters.
}
\description{
Gamma Distribution
}
