% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Norm.R
\name{Norm}
\alias{Norm}
\alias{d,Norm-method}
\alias{p,Norm-method}
\alias{q2,Norm-method}
\alias{r,Norm-method}
\alias{mean,Norm-method}
\alias{median,Norm-method}
\alias{mode,Norm-method}
\alias{var,Norm-method}
\alias{sd,Norm-method}
\alias{skew,Norm-method}
\alias{kurt,Norm-method}
\alias{entro,Norm-method}
\alias{finf,Norm-method}
\alias{ll,numeric,numeric,Norm-method}
\alias{mle,numeric,Norm-method}
\alias{me,numeric,Norm-method}
\alias{avar_mle,Norm-method}
\alias{avar_me,Norm-method}
\title{Normal Distribution}
\usage{
Norm(mean = 0, sd = 1)

\S4method{d}{Norm}(x)

\S4method{p}{Norm}(x)

\S4method{q2}{Norm}(x)

\S4method{r}{Norm}(x)

\S4method{mean}{Norm}(x)

\S4method{median}{Norm}(x)

\S4method{mode}{Norm}(x)

\S4method{var}{Norm}(x)

\S4method{sd}{Norm}(x)

\S4method{skew}{Norm}(x)

\S4method{kurt}{Norm}(x)

\S4method{entro}{Norm}(x)

\S4method{finf}{Norm}(x)

\S4method{ll}{numeric,numeric,Norm}(x, prm, distr)

\S4method{mle}{numeric,Norm}(x, distr)

\S4method{me}{numeric,Norm}(x, distr)

\S4method{avar_mle}{Norm}(distr)

\S4method{avar_me}{Norm}(distr)
}
\arguments{
\item{mean, sd}{numeric. The distribution parameters.}

\item{x}{an object of class \code{Norm}. If the function also has a \code{distr}
argument, \code{x} is a numeric vector, a sample of observations.}

\item{prm}{numeric. A vector including the distribution parameters.}

\item{distr}{an object of class \code{Norm}.}
}
\description{
Normal Distribution
}
