% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Pois.R
\name{Pois}
\alias{Pois}
\alias{d,Pois-method}
\alias{p,Pois-method}
\alias{q2,Pois-method}
\alias{r,Pois-method}
\alias{mean,Pois-method}
\alias{var,Pois-method}
\alias{sd,Pois-method}
\alias{skew,Pois-method}
\alias{kurt,Pois-method}
\alias{finf,Pois-method}
\alias{ll,numeric,numeric,Pois-method}
\alias{mle,numeric,Pois-method}
\alias{me,numeric,Pois-method}
\alias{avar_mle,Pois-method}
\alias{avar_me,Pois-method}
\title{Poisson Distribution}
\usage{
Pois(lambda = 1)

\S4method{d}{Pois}(x)

\S4method{p}{Pois}(x)

\S4method{q2}{Pois}(x)

\S4method{r}{Pois}(x)

\S4method{mean}{Pois}(x)

\S4method{var}{Pois}(x)

\S4method{sd}{Pois}(x)

\S4method{skew}{Pois}(x)

\S4method{kurt}{Pois}(x)

\S4method{finf}{Pois}(x)

\S4method{ll}{numeric,numeric,Pois}(x, prm, distr)

\S4method{mle}{numeric,Pois}(x, distr)

\S4method{me}{numeric,Pois}(x, distr)

\S4method{avar_mle}{Pois}(distr)

\S4method{avar_me}{Pois}(distr)
}
\arguments{
\item{lambda}{numeric. The distribution parameters.}

\item{x}{an object of class \code{Pois}. If the function also has a \code{distr}
argument, \code{x} is a numeric vector, a sample of observations.}

\item{prm}{numeric. A vector including the distribution parameters.}

\item{distr}{an object of class \code{Pois}.}
}
\value{
The dpqr family of functions return the evaluated density, cumulative
probability, quantile, and random sample, respectively.
The moments family of functions return the appropriate theoretical moment,
as calculated by the distribution true parameters.
The ll function returns the evaluated log-likelihood, given a sample and the
theoretical parameters.
The estim family of functions return the estimated parameters of the
distribution, given a sample.
The avar family of functions return the asymptotic variance or variance -
covariance matrix (if there are two or more parameters) of the corresponding
estimation method.
Calculus performed on Distribution objects returns a Distribution object of
the appropriate class and with the appropriate parameters.
}
\description{
Poisson Distribution
}
