% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same.R
\name{same}
\alias{same}
\alias{same,ANY,character-method}
\alias{same,numeric,Beta-method}
\alias{same,matrix,Beta-method}
\alias{same,numeric,Gammad-method}
\alias{same,matrix,Gammad-method}
\alias{same,matrix,Dirichlet-method}
\alias{same,array,MGamma-method}
\title{Score-Adjusted Moment Estimator}
\usage{
same(x, distr, ...)

\S4method{same}{ANY,character}(x, distr, ...)

\S4method{same}{numeric,Beta}(x, distr)

\S4method{same}{matrix,Beta}(x, distr)

\S4method{same}{numeric,Gammad}(x, distr)

\S4method{same}{matrix,Gammad}(x, distr)

\S4method{same}{matrix,Dirichlet}(x, distr)

\S4method{same}{array,MGamma}(x, distr)
}
\arguments{
\item{x}{numeric. A sample under estimation.}

\item{distr}{A subclass of \code{Distribution}. The distribution of interest.}

\item{...}{extra arguments.}
}
\value{
numeric. The estimator produced by the sample.
}
\description{
Score-Adjusted Moment Estimator
}
