% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same.R
\name{same}
\alias{same}
\alias{same,ANY,character-method}
\alias{same,numeric,Beta-method}
\alias{same,matrix,Beta-method}
\alias{same,numeric,Gammad-method}
\alias{same,matrix,Gammad-method}
\alias{same,matrix,Dirichlet-method}
\alias{same,matrix,MGamma-method}
\title{Score-Adjusted Moment Estimator}
\usage{
same(x, distr, ...)

\S4method{same}{ANY,character}(x, distr, ...)

\S4method{same}{numeric,Beta}(x, distr)

\S4method{same}{matrix,Beta}(x, distr)

\S4method{same}{numeric,Gammad}(x, distr)

\S4method{same}{matrix,Gammad}(x, distr)

\S4method{same}{matrix,Dirichlet}(x, distr)

\S4method{same}{matrix,MGamma}(x, distr)
}
\arguments{
\item{x}{numeric. A sample under estimation.}

\item{distr}{A subclass of \code{Distribution}. The distribution family assumed.}

\item{...}{extra arguments.}
}
\value{
numeric. The estimator produced by the sample.
}
\description{
Calculates the SAMΕ of a sample under the assumption the observations are
independent and identically distributed (iid) according to a specified
family of distributions.
}
\examples{
\dontrun{
# Distribution
D <- Beta(shape1 = 1, shape2 = 1.5)

# Simulation
set.seed(2)
x <- r(D)(50)

# Estimation
me(x, "Beta")
same(x, "Beta")
mle(x, "Beta")

# Asymptotic Covariance Matrix
acov_me(D)
acov_same(D)
acov_mle(D)
}
}
