% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Stud.R
\name{Stud}
\alias{Stud}
\alias{d,Stud-method}
\alias{p,Stud-method}
\alias{q2,Stud-method}
\alias{r,Stud-method}
\alias{mean,Stud-method}
\alias{median,Stud-method}
\alias{mode,Stud-method}
\alias{var,Stud-method}
\alias{sd,Stud-method}
\alias{skew,Stud-method}
\alias{kurt,Stud-method}
\title{Student Distribution}
\usage{
Stud(df = 1, ncp = 0)

\S4method{d}{Stud}(x)

\S4method{p}{Stud}(x)

\S4method{q2}{Stud}(x)

\S4method{r}{Stud}(x)

\S4method{mean}{Stud}(x)

\S4method{median}{Stud}(x)

\S4method{mode}{Stud}(x)

\S4method{var}{Stud}(x)

\S4method{sd}{Stud}(x)

\S4method{skew}{Stud}(x)

\S4method{kurt}{Stud}(x)
}
\arguments{
\item{df, ncp}{numeric. The distribution parameters.}

\item{x}{an object of class \code{Stud}. If the function also has a \code{distr}
argument, \code{x} is a numeric vector, a sample of observations.}
}
\description{
Student Distribution
}
