% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Unif.R
\name{Unif}
\alias{Unif}
\alias{d,Unif-method}
\alias{p,Unif-method}
\alias{q2,Unif-method}
\alias{r,Unif-method}
\alias{mean,Unif-method}
\alias{var,Unif-method}
\alias{sd,Unif-method}
\alias{skew,Unif-method}
\alias{kurt,Unif-method}
\alias{entro,Unif-method}
\alias{ll,numeric,numeric,Unif-method}
\alias{mle,numeric,Unif-method}
\alias{me,numeric,Unif-method}
\title{Uniform Distribution}
\usage{
Unif(min = 0, max = 1)

\S4method{d}{Unif}(x)

\S4method{p}{Unif}(x)

\S4method{q2}{Unif}(x)

\S4method{r}{Unif}(x)

\S4method{mean}{Unif}(x)

\S4method{var}{Unif}(x)

\S4method{sd}{Unif}(x)

\S4method{skew}{Unif}(x)

\S4method{kurt}{Unif}(x)

\S4method{entro}{Unif}(x)

\S4method{ll}{numeric,numeric,Unif}(x, prm, distr)

\S4method{mle}{numeric,Unif}(x, distr)

\S4method{me}{numeric,Unif}(x, distr)
}
\arguments{
\item{min, max}{numeric. The distribution parameters.}

\item{x}{an object of class \code{Unif}. If the function also has a \code{distr}
argument, \code{x} is a numeric vector, a sample of observations.}

\item{prm}{numeric. A vector including the distribution parameters.}

\item{distr}{an object of class \code{Unif}.}
}
\description{
Uniform Distribution
}
